Crate btls

btls: Relatively Bare-bones TLS library.

This crate is a relatively bare-bones TLS library.

Only pretty basic TLS features are supported, including:

TLS 1.2

AES-128-GCM, AES-256-GCM and Chacha20-Poly1305

ECDHE with NSA P-256, NSA P-384, X25519 and X448.

Support for RSA (PKCS#1 and PSS) and ECDSA signatures.

(Experimential) Ed25519 and Ed448 signatures.

Server support for external signers (RSA, RSA-PSS, ECDSA and EdDSA).

Server support for external certificate selection

OSCP stapling and verification

CT stapling and verification

Automatic certificate selection with interlocking (server only)

Certificate validation (client only)

TLS Exporters

SNI support

ALPN support

RPK support

EMS support (including option to require EMS).

TLS features 5 (must-staple), 18 (must-ct) and 23 (must-ems).

Pinning and TLSA (client only).

To create a client TLS session, create a ClientConfiguration, set up the parameters and then call .make_session() to get the ClientSession object, which is the client session.

To create a server TLS session, create a ServerConfiguration, set up the parameters and then call .make_session() to get the ServerSession object, which is the server session.

To see operations one can do on sessions (both client and server), see the Session trait.

Optional cargo features:

gnutls-ecdsa: Use GnuTLS library for ECDSA signing (needs GnuTLS installed). Otherwise internal ECDSA signing is not supported.

enable-key-debug: Enable DEBUG_CRYPTO_CALCS logging (only helpful for diagnosing interop failures due to keys somehow being computed wrong).

Modules

client_certificates
Types related to certificate handling on client

error
Errors that occur during various operations.

features
This module contains definitions of feature flags.

futures
Simple implementation of futures.

jose_jws
JOSE JWS signing.

logging
Types related to logging

server_certificates
Types related to server certificate store.

server_keys
Types related to server keys.

test
Internal testing stuff.

Structs

ClientConfiguration
Client configuration.

ClientSession
A client TLS session.

ServerConfiguration
Server configuration.

ServerSession
A server TLS session.

TlsConnectionInfo
Information about connection.

TxIrqFlag
A TLS Session TX IRQ flag.

Enums

AbortReason
Reason TLS connection got aborted.

IoErrorOrTlsFailure
IoError or TlsFailure

Traits

RxAppdataHandler
A TLS session appdata RX handler.

Session
A TLS session.

TxIrqHandler
A TLS Session TX IRQ handler.

Functions

unwrap_tls_failure
Take a IoResult error, and return embedded TlsFailure if any or return the original.