Struct btls::client_certificates::HostSpecificPin

pub struct HostSpecificPin {
    pub ca_flag: Option<bool>,
    pub pin: bool,
    pub ta: bool,
    pub spki: bool,
    pub data: DataRepresentation,
}

Host-specific certificate pin / trust anchor.

This structure contains a host-specific certificate pin or trust anchor.

It can be used e.g. for implemeting DANE TLSA (RFC6698) support or HPKP (RFC 7469).

Notes:

Putting a pin that is pinned (or no pin is actually pinned) and trust anchor on EE SPKI will automatically light up RPK support.

There are currently following limits for matching:

Built-in roots can only be matched by SPKI.

Pins can only match certificates in chain, or built-in roots. This applies even if data representation is uncompressed and whole certificate matching is used.

Fields
ca_flag: Option<bool>

CA or EE flag. If true, matches only CAs, if false, matches only EEs, if none, matches both.

pin: bool

Pin flag. If set, the entry acts as a pin.

ta: bool

Trust Anchor flag. If set, the entry acts as a trust anchor.

spki: bool

SPKI flag. If set, entry matches SPKI, otherwise it matches entiere certificate.

data: DataRepresentation

The data to match against.

Methods

impl HostSpecificPin

fn trust_server_key_by_sha256(sha256: &[u8; 32],
whole_cert: bool,
issuer: bool)
-> HostSpecificPin

Trust server key with specified SHA-256.

Parameters:

sha256: The SHA-256 of the key.

whole_cert: If true, hash the whole cert instead of just the key (the SPKI).

issuer: If true, accept CA certificates instead of end-entity certificates.

Returns:

The pin object.

Notes:

The object is marked as trust anchor but not pin, so don't mix this with pinning.

The certificate has to be sent by the server for this to work properly (only really matters for CA certificates, since server keys are always sent).

Examples:

This might be handy for those testing certificates:

let config = ClientConfiguration::new();
let mut connection = config.make_session_pinned(TEST_SERVER_NAME, &[trust_server_key_by_sha256(TEST_SERVER_KEYHASH, false, false)]);

This lets you to connect to server TEST_SERVER_NAME having self-signed certificate with key SHA-256 fingerprint of TEST_SERVER_KEYHASH.

fn from_tlsa_raw(raw_tlsa: &[u8]) -> Result<HostSpecificPin, ()>

Create a host specific pin from raw DNS TLSA data.

Parameters:

raw_tlsa: The raw RRDATA of TLSA record.

Returns:

The pin object.

Failures:

If the data is malformed, fails with ().