Struct btls::client_certificates::TrustAnchor

pub struct TrustAnchor { /* fields omitted */ }

A trust anchor.

This structure represents a trust anchor to evaluate trust against.

There are three basic ways of creating this structure:

Pass the subject (as X.509 Name) and public key (as X.509 SubjectPublicKeyInfo), along with optional name constraints to TrustAnchor::from_name_spki().

Pass certificate data in DER form to TrustAnchor::from_certificate().

Pass filename of file containg the certificate data in DER form to TrustAnchor::from_certificate_file().

Methods

impl TrustAnchor

fn from_name_spki(subject: &[u8],
spki: &[u8],
allow_names: &[&str],
disallow_names: &[&str])
-> Result<TrustAnchor, String>

Create new trust anchor from subject name and SPKI.

Parameters:

subject: The subject name, as X.509 Name structure.

spki: The X.509 SubjectPublicKeyInfo structure for the signature key.

allowed_names: Optional name constraints (if empty, all names are allowed).

disallowed_names: Optional name constraints.

Return value:

The new trust anchor.

Failures:

If the data is just too mangled, fails with descriptive error.

fn from_certificate<R: IoRead>(certificate: &mut R)
-> Result<TrustAnchor, String>

Create a trust anchor from DER-formatted certificate.

Parameters:

certificate: The certificate to use as trust anchor.

Return value:

The new trust anchor.

Failures:

If the data is just too mangled, fails with descriptive error.

fn from_certificate_file<P: AsRef<Path>>(certificate: P)
-> Result<TrustAnchor, String>

Create a trust anchor from DER-formatted certificate file.

Parameters:

certificate: The filename of certificate to use as trust anchor.

Return value:

The new trust anchor.

Failures:

On error, returns descriptive error.