Enum btls::error::CertificateError

pub enum CertificateError {
    Toplevel(CertificateToplevelError),
    Version(CertificateVersionError),
    Serial(Asn1Error),
    SignatureAlgorithm(Asn1Error),
    Issuer(Asn1Error),
    Validity(CertificateValidityError),
    Subject(Asn1Error),
    SubjectPublicKeyInfo(Asn1Error),
    SkipElements(Asn1Error),
    NoExtensions,
    Extensions(Asn1Error),
    ExtensionEntry(usize, CertificateExtensionError),
    AnyKeyUsageInEeCert,
    NameConstraintsInEeCert,
    NotCaCanSignCertificates,
    NonIntegralSignatureOctets,
    // some variants omitted
}

Error in parsing certificate.

Variants
Toplevel(CertificateToplevelError)

Error in reading toplevel.

This error occurs if anything goes wrong in reading the top level of the certificate, which includes the topmost SEQUENCE and the fields to-be-signed data, signature algorithm and signature elements within that sequence.

The parameter is structure giving more detailed reading error.

Version(CertificateVersionError)

Error in parsing certificate version.

This error occurs if anything goes wrong in parsing the certificate version, or if the certificate version is not supported.

The parameter is structure giving more detailed parsing error.

Serial(Asn1Error)

Error in reading certificate serial.

This error occurs if anything goes wrong in reading the certificate serial number.

The parameter is structure giving more detailed reading error.

SignatureAlgorithm(Asn1Error)

Error in reading certificate signature algorithm.

This error occurs if anything goes wrong in reading the certificate signature algorithm.

The parameter is structure giving more detailed reading error.

Issuer(Asn1Error)

Error in reading certificate issuer.

This error occurs if anything goes wrong in reading the certificate issuer.

The parameter is structure giving more detailed reading error.

Validity(CertificateValidityError)

Error in parsing certificate validity.

This error occurs if anything goes wrong in parsing the certificate validity limits.

The parameter is structure giving more detailed parsing error.

Subject(Asn1Error)

Error in reading certificate subject.

This error occurs if anything goes wrong in reading the certificate subject.

The parameter is structure giving more detailed reading error.

SubjectPublicKeyInfo(Asn1Error)

Error in reading certificate SubjectPublicKeyInfo.

This error occurs if anything goes wrong in reading the certificate SubjectPublicKeyInfo.

The parameter is structure giving more detailed reading error.

SkipElements(Asn1Error)

Error in reading skipped elements before extensions.

This error occurs if anything goes wrong in skipping various elements between SubjectPublicKeyInfo and extensions..

The parameter is structure giving more detailed reading error.

NoExtensions

Certificate has no extensions.

This error occurs if the certificate has no extensions.

Extensions(Asn1Error)

Error in reading certificate extensions.

This error occurs if the top-level SEQUENCE of the extensions structure can not be read.

The parameter is structure giving more detailed reading error.

ExtensionEntry(usize, CertificateExtensionError)

Error in parsing certificate extension.

This error occurs if extension structure can not be parsed.

The 1st parameter is number of extension (starting from 0) that triggered the parse error.

The 2nd parameter is more detailed parsing errorr.

AnyKeyUsageInEeCert

AnyKeyUsage used in EE certificate.

This error occurs if EE certificate has AnyKeyUsage EKU asserted.

NameConstraintsInEeCert

Name constraints used in EE certificate.

This error occurs if EE certificate has name constraints present.

NotCaCanSignCertificates

Certificate is not CA but can sign certificates?

This error occurs if certificate has BasicConstraints CA=False, but still asserts the CertSign KeyUsage.

NonIntegralSignatureOctets

The number of signature octets is not integral.

This error occurs if signature length is not an integral multiple of an octet.