Enum btls::error::CertificateExtensionError

pub enum CertificateExtensionError {
    Toplevel(Asn1Error),
    Oid(Asn1Error),
    Critical(String, Asn1Error),
    Value(String, Asn1Error),
    Eku(Asn1Error),
    EkuEntry(usize, Asn1Error),
    BasicConstraints(Asn1Error),
    BasicConstraintsCa(Asn1Error),
    BasicConstraintsPathlen(Asn1Error),
    KeyUsage(Asn1Error),
    TlsFeature(Asn1Error),
    TlsFeatureEntry(usize, Asn1Error),
    Sans(NameListError),
    NameConstraints(NameConstraintsError),
    Scts(Asn1Error),
    SctsListOverflow,
    SctsEntryOverflow,
    UnrecognizedCritical(String),
    // some variants omitted
}

Error parsing certificate extension.

Variants
Toplevel(Asn1Error)

Error reading the extension toplevel.

This error occurs if the topmost SEQUENCE in the extension can not be read.

The parameter is structure giving more detailed reading error.

Oid(Asn1Error)

Error reading the extension OID.

This error occurs if the OID of the extension can not be read.

The parameter is structure giving more detailed reading error.

Critical(String, Asn1Error)

Error reading the extension critical flag

This error occurs if the critical flag of extension can not be read.

The 1st parameter is string representation of the OID of the extension.

The 2nd parameter is more detailed reading error.

Value(String, Asn1Error)

Error reading the extension value.

This error occurs if the value of extension can not be read.

The 1st parameter is string representation of the OID of the extension.

The 2nd parameter is more detailed reading error.

Eku(Asn1Error)

Error reading the ExtendedKeyUsage list.

This error occurs if toplevel of ExtendedKeyUsage extension can not be read.

The parameter is structure giving more detailed reading error.

EkuEntry(usize, Asn1Error)

Error reading ExtendedKeyUsage list entry.

This error occurs if the value in ExtendedKeyUsage list can not be read.

The 1st parameter is the 0-based number of EKU entry that can't be read.

The 2nd parameter is more detailed reading error.

BasicConstraints(Asn1Error)

Error reading BasicConstraints.

This error occurs if toplevel of BasicConstraints extension can not be read.

The parameter is structure giving more detailed reading error.

BasicConstraintsCa(Asn1Error)

Error reading BasicConstraints CA flag.

This error occurs if the CA flag in BasicConstraints extension can not be read.

The parameter is structure giving more detailed reading error.

BasicConstraintsPathlen(Asn1Error)

Error reading BasicConstraints pathlen limit.

This error occurs if the pathlen limit in BasicConstraints extension can not be read.

The parameter is structure giving more detailed reading error.

KeyUsage(Asn1Error)

Error reading KeyUsage.

This error occurs if toplevel of KeyUsage extension can not be read.

The parameter is structure giving more detailed reading error.

TlsFeature(Asn1Error)

Error reading TlsFeature list.

This error occurs if toplevel of TlsFeature extension can not be read.

The parameter is structure giving more detailed reading error.

TlsFeatureEntry(usize, Asn1Error)

Error reading TlsFeature list entry.

This error occurs if the value in TlsFeature list can not be read.

The 1st parameter is the 0-based number of TlsFeature entry that can't be read.

The 2nd parameter is more detailed reading error.

Sans(NameListError)

Error parsing SANs.

This error occurs if SubjectAlternativeName extension can not be parsed.

The parameter is structure giving more detailed parsing error.

NameConstraints(NameConstraintsError)

Error parsing name constraints.

This error occurs if NameConstraints extension can not be parsed.

The parameter is structure giving more detailed parsing error.

Scts(Asn1Error)

Error reading SCTs.

This error occurs if toplevel of embedded SCTs extension can not be read.

The parameter is structure giving more detailed reading error.

SctsListOverflow

SCT list overflows the extension value.

This error occurs if the list of SCTs inside embedded SCTs extension overflows the parent extension.

SctsEntryOverflow

SCT entry overflows the SCT list.

This error occurs if one of the SCTs inside embedded SCTs list overflows the parent list.

UnrecognizedCritical(String)

Unrecognized critical extension.

This error occurs if unknown extension has the critical flag set.

The parameter structure gives the OID of the offending extension.