Enum btls::error::HandshakeState

pub enum HandshakeState {
    ClientHello,
    HelloRestartRequest,
    RetryClientHello,
    ServerHello,
    ServerHelloWaitAgain,
    Tls12Certificate,
    Tls12ServerKeyExchange,
    Tls12ServerHelloDone,
    Tls12ClientCertificate,
    Tls12ClientKeyExchange,
    Tls12ClientChangeCipherSpec,
    Tls12ClientFinished,
    Tls12ServerChangeCipherSpec,
    Tls12ServerFinished,
    Tls12Showtime,
    Tls13EncryptedExtensions,
    Tls13Certificate,
    Tls13CertificateVerify,
    Tls13ServerFinished,
    Tls13ClientCertificate,
    Tls13ClientFinished,
    Tls13Showtime,
    Error,
    // some variants omitted
}

The state of handshake.

Variants
ClientHello

Next, client sends ClientHello to server.

HelloRestartRequest

Next, the server will send HelloRestartRequest to client (this state is only valid on server and for TLS 1.3).

RetryClientHello

Next, client retries sending ClientHello to server (TLS 1.3 only).

ServerHello

Next, server sends ServerHello to client.

ServerHelloWaitAgain

The client is waiting for ServerHello in response to retry from the server (TLS 1.3 only, only valid on client).

Tls12Certificate

Next, the server will send TLS 1.2 Certificate message (TLS 1.2 only).

Tls12ServerKeyExchange

Next, the server will send TLS 1.2 ServerKeyExchange message (TLS 1.2 only). Also, preceeding that message may be CertificateStatus message from the server.

Tls12ServerHelloDone

Next, the server will send TLS 1.2 ServerHelloDone message (TLS 1.2 only). Also, preceeding that message may be CertificateRequest message from the server (on client only).

Tls12ClientCertificate

Next, the client will send TLS 1.2 Certificate message NAKing authentication to the server (TLS 1.2 only, only valid on client).

Tls12ClientKeyExchange

Next, the client will send TLS 1.2 ClientKeyExchange message to the server (TLS 1.2 only).

Tls12ClientChangeCipherSpec

Next, the client will send TLS 1.2 ChangeCipherSpec message to the server (TLS 1.2 only).

Tls12ClientFinished

Next, the client will send TLS 1.2 Finished message to the server (TLS 1.2 only).

Tls12ServerChangeCipherSpec

Next, the server will send TLS 1.2 ChangeCipherSpec message to the client (TLS 1.2 only).

Tls12ServerFinished

Next, the server will send TLS 1.2 Finished message to the client (TLS 1.2 only).

Tls12Showtime

TLS 1.2 connection is up and running.

Tls13EncryptedExtensions

Next, the server will send TLS 1.3 EncryptedExtensions message to the client (TLS 1.3 only).

Tls13Certificate

Next, the server will send TLS 1.3 Certificate message to the client (TLS 1.3 only).

Tls13CertificateVerify

Next, the server will send TLS 1.3 CertificateVerify message to the client (TLS 1.3 only).

Tls13ServerFinished

Next, the server will send TLS 1.3 Finished message to the client (TLS 1.3 only).

Tls13ClientCertificate

Next, the client will send TLS 1.3 Certificate message, NAKing authentication to the server (TLS 1.3 only, only valid on client).

Tls13ClientFinished

Next, the client will send TLS 1.3 Finished to the server (TLS 1.3 only)

Tls13Showtime

TLS 1.3 connection is up and running.

Error

Handshake has failed.

Methods

impl HandshakeState

fn get_direction(&self) -> HandshakeDirection

Get the direction of next handshake message.

Parameters:

self: The state to get information about.

Returns:

The direction of next handshake message.