Enum btls::error::TlsFailure

pub enum TlsFailure {
    UnsupportedVersion(u16),
    ServerHelloNotAlignedTls13,
    ServerHelloDoneNotEmpty(usize),
    FinishedNotAligned,
    InterlockSignatureMissing,
    KeyUpdateNotAligned,
    KeyUpdateBadLength(usize),
    EmptyHandshakeMessage,
    CcsInMiddleOfHandshakeMessage,
    InvalidCcs,
    UnexpectedRtype(u8),
    UnexpectedHandshakeMessage(u8, HandshakeState),
    UnexpectedCcs(HandshakeState),
    NoNontrivialKex,
    MaxPrfSizeTooSmall,
    HsMsgBuffer(HsBufferError),
    ClientHelloTooLong(usize, usize),
    AcmeValidationDone,
    TxSerializationOverflow,
    TxTooBig(String),
    TxSerializeOverflow,
    NoCertSelectorFuture,
    CertSelectorFutureError,
    NoCertificateAvailable(String),
    EmptyCertChain(String),
    EmptyOcsp,
    NoSignatureFuture,
    SignatureFutureError,
    CantSignTbs,
    CantRestoreSke,
    CantObtainSignatureTbs,
    SctListEmpty,
    ServerNameExpectList,
    ServerNameExpectEntryType,
    ServerNameExpectEntryContent,
    EmptyHostname,
    TooLongHostname,
    BogusHostname,
    AlpnExpectedList,
    AlpnExpectedEntry,
    EmptyAlpn,
    BogusAlpn,
    AlpnNoOverlap,
    SupGrpExpectedList,
    SupGrpExpectedGroup,
    SigAlgoExpectedList,
    SigAlgoExpectedAlgorithm,
    KeyShareExpectedList,
    KeyShareExpectedGroup,
    KeyShareExpectedContent,
    CkeCurveError(u16, KeyAgreementError),
    ChBogusTlsVersion(u16),
    KeyShareCurveError(u16, KeyAgreementError),
    DhKeypairGenError(u16, KeyAgreementError),
    GenChalCertFailed(AcmeCertificateCreateError),
    KeyShareDupGroup(u16),
    InterlockOcsp(OcspValidationError),
    CantGetOwnDhPubKey(KeyAgreementError),
    NoMutualCiphersuites,
    NoMutualGroups,
    NoKeyshareInTls13,
    ChExpectedVersion,
    ChExpectedRandom,
    ChExpectedSessionId,
    ChExpectedCiphersuites,
    ChExpectedCompressionMethods,
    ChExpectedExtensions,
    ChExpectedExtensionKind,
    ChExpectedExtensionPayload(u16),
    ChUnsupportedClientVersion(u16),
    ChBogusCompressionMethods(Vec<u8>),
    ChDuplicateExtension(u16),
    ChJunkAfterExtension(u16),
    ChSsl3NotSupported,
    ChTls10NotSupported,
    ChTls11NotSupported,
    SupVerJunkAfterList,
    SupVerExpectedVersionList,
    SupVerExpectedVersion,
    NoMutualTlsVersion,
    ChExpectedCiphersuite,
    InapporiateFallback,
    Ciphersuite0,
    NoMutualCompression,
    ChJunkAfterEnd,
    NoMutualCertificateType,
    SCertTypeExpectedList,
    SCertTypeExpectedFormat,
    KeyShareNotAllowedInTls12,
    ChRenegotiationAttackDetected,
    ShRenegotiationAttackDetected,
    Tls13RequresKeyShare,
    KeySharesNotSubsetSupGrp,
    ClientVulernableTHS,
    ClientVulernableRenego,
    ChDoubleRestart,
    CkeExpectedPublic,
    FinishedMacFailed,
    CantExtractSpki(CertificateError),
    InterlockCertificate(CertificateValidationError),
    NullExtractor,
    BadTls12MsSize,
    KeyblockTooSmall,
    MasterblockTooSmall,
    CantCreateProtector,
    CantCreateDeprotector,
    TextOutputTooSmall,
    ConcatenationNotUtf8,
    InterlockNontrivalKex,
    WrongClientRandomLength(usize),
    WrongServerRandomLength(usize),
    TlsExtractorsRequireEmsOrTls13,
    Tls13ExporterLabelTooLong,
    Tls13ExporterContextTooLong,
    Tls13ExporterOutputTooLong,
    CantLockSession,
    CvSignatureFailed,
    SctRequiredForCert,
    CantParseIssuerCertificate(CertificateError),
    CkeWrongNumberOfShares(usize),
    HrrExpectedVersion,
    HrrBogusTlsVersion(u16),
    HrrTlsVersionUnsupported(u16),
    HrrExpectedExtensionKind,
    HrrExpectedExtensionPayload(u16),
    HrrDuplicateExtension(u16),
    HrrExpectedKeyShare,
    HrrExpectedCookie,
    HrrUnsupportedExtension(u16, u16),
    HrrExpectedExtensions,
    HrrExpectedEnd,
    HrrBogusRestart,
    HrrAlreadyHadShare(u16),
    HrrBogusCurve(u16),
    AlpnExpectedLength1,
    AlpnExpectedLength2,
    AlpnEmptyProtocol,
    AlpnRawLengthTooBig,
    AlpnExpectedLen1Val(usize, usize),
    AlpnExpectedLen2Val(usize, usize),
    AlpnNonUtf8Protocol,
    AlpnBogusProtocol,
    SKeyShareExpectedGroup,
    SKeyShareExpectedKex,
    KeyShareGroupDisagreement(u16, u16),
    KeyShareBogusGroup(u16),
    ShExpectedVersion,
    ShSsl3NotSupported,
    ShTls10NotSupported,
    ShTls11NotSupported,
    ShBogusVersion(u16),
    ShUnsupportedTlsVersion(u16),
    ShExpectedRandom,
    ShBadRandomLength,
    Tls13DowngradeAttack,
    ShExpectedSessionId,
    ShBogusCiphersuite(u16),
    ShExpectedCiphersuite,
    ShBadTls12Cipher(u16),
    ShBadTls13Cipher(u16),
    ShExpectedCompression,
    ShBogusCompression(u8),
    ShExpectedExtensionKind,
    ShExpectedExtensionPayload(u16),
    ShDuplicateExtension(u16),
    SctExpectedList,
    SctExpectedEntry,
    SCertTypeExpectedFormatE,
    ShUnsupportedExtension(u16, u16),
    ShJunkAfterExtension(u16),
    ShExpectedExtensions,
    ShExpectedEnd,
    Tls13NoKeyShare,
    ServerVulernableTHS,
    ServerVulernableRenego,
    CertRequiresOcsp,
    CertRequiresSct,
    SkeExpectedCurveType,
    SkeBogusCurveType(u8),
    SkeExpectedNamedcurve,
    SkeBogusCurve(u16),
    SkeExpectedPublic,
    SkeCurveError(u16, KeyAgreementError),
    SkeExpectedAlgorithm,
    SkeBogusSignatureAlgo(u16),
    SkeExpectedSignature,
    SkeExpectedEnd,
    SkeSignatureFailed,
    ServerKeyNotWhitelisted,
    EECertParseError(CertificateError),
    SCertTypeBogusType(u8),
    ServerChainEmpty,
    EeCertificateEmpty,
    CsBogusOcspType(u8),
    CantExtractSct12(OcspValidationError),
    OcspIncompleteCertificateChain,
    InvalidOcsp(OcspValidationError),
    CrExpectedContext,
    CrContextMustBeEmpty,
    CantExtractSct13(OcspValidationError),
    CertificateValidationFailed(CertificateValidationError),
    CertExpectCertList,
    CertExpectIntermeidate,
    CertIntermediateEmpty,
    CertExpectedContext,
    CertExpectExtensionKind,
    CertExpectExtionPayload(u16),
    CertDuplicateExtension(u16),
    OcspExpectedType,
    OcspExpectedContent,
    SCertTypeInNotFirstCert,
    CertUnsupportedExtension(u16),
    CertJunkAfterExtension(u16),
    CertExpectedExtensions,
    CvExpectedAlgorithm,
    CvBogusSignatureAlgo(u16),
    CvExpectedSignature,
    CvExpectedEnd,
    CsExpectedStatusType,
    CsExpectedResponse,
    CsEmptyResponse,
    EeExpectedExtensionKind,
    EeExpectedExtensionPayload(u16),
    EeDuplicateExtension(u16),
    EeUnsupportedExtension(u16),
    EeJunkAfterExtension(u16),
    EeExpectedExtensions,
    EeExpectedEnd,
    DeprotectionFailed,
    AppdataBeforeFinished,
    AppdataAfterEof,
    RecvAppdataRQueueEof,
    MaxRecordScratchOverflow,
    TxHandshakeMsgOver16MB(usize),
    FragmentTooBig(usize),
    InvalidFragRtype(u8),
    PacketLengthOutOfWhack(usize),
    RecordTooBig(usize),
    NoPacketToProcess,
    ShortreadRecordHeader,
    RxRecordTooBig(usize),
    PacketLengthToSendOutOfWhack(usize),
    ProtectingPaceketFailed,
    TxQueueWriteFailed(QueueWriteError),
    NormalClose,
    ProtectingPaceketFailedZ,
    ReadDiscardFailed(QueueReadError),
    ValidationHandshakeDone,
    NoRecordExpected,
    UnknownIoException,
    TxQueueEof,
    TxQueueDiscard(QueueReadError),
    RxQueueEof,
    WQueueDataNotRead,
    StreamReadOverflow,
    NoTrustAnchors,
    CantLockAcmeChallenges,
    InternalConnection,
    TlsSessionError,
    IntErrNoIssuerSerial,
    SctValidationFailed(SctError),
    // some variants omitted
}

Failure in TLS library.

This lists all the failures that can happen in TLS handshake and operation.

Variants
UnsupportedVersion(u16)

Unsupported TLS version.

ServerHelloNotAlignedTls13

ServerHello does not end last fragment it is in.

ServerHelloDoneNotEmpty(usize)FinishedNotAlignedInterlockSignatureMissingKeyUpdateNotAlignedKeyUpdateBadLength(usize)EmptyHandshakeMessageCcsInMiddleOfHandshakeMessageInvalidCcsUnexpectedRtype(u8)UnexpectedHandshakeMessage(u8, HandshakeState)UnexpectedCcs(HandshakeState)NoNontrivialKexMaxPrfSizeTooSmallHsMsgBuffer(HsBufferError)ClientHelloTooLong(usize, usize)AcmeValidationDoneTxSerializationOverflowTxTooBig(String)TxSerializeOverflowNoCertSelectorFutureCertSelectorFutureErrorNoCertificateAvailable(String)EmptyCertChain(String)EmptyOcspNoSignatureFutureSignatureFutureErrorCantSignTbsCantRestoreSkeCantObtainSignatureTbsSctListEmptyServerNameExpectListServerNameExpectEntryTypeServerNameExpectEntryContentEmptyHostnameTooLongHostnameBogusHostnameAlpnExpectedListAlpnExpectedEntryEmptyAlpnBogusAlpnAlpnNoOverlapSupGrpExpectedListSupGrpExpectedGroupSigAlgoExpectedListSigAlgoExpectedAlgorithmKeyShareExpectedListKeyShareExpectedGroupKeyShareExpectedContentCkeCurveError(u16, KeyAgreementError)ChBogusTlsVersion(u16)KeyShareCurveError(u16, KeyAgreementError)DhKeypairGenError(u16, KeyAgreementError)GenChalCertFailed(AcmeCertificateCreateError)KeyShareDupGroup(u16)InterlockOcsp(OcspValidationError)CantGetOwnDhPubKey(KeyAgreementError)NoMutualCiphersuitesNoMutualGroupsNoKeyshareInTls13ChExpectedVersionChExpectedRandomChExpectedSessionIdChExpectedCiphersuitesChExpectedCompressionMethodsChExpectedExtensionsChExpectedExtensionKindChExpectedExtensionPayload(u16)ChUnsupportedClientVersion(u16)ChBogusCompressionMethods(Vec<u8>)ChDuplicateExtension(u16)ChJunkAfterExtension(u16)ChSsl3NotSupportedChTls10NotSupportedChTls11NotSupportedSupVerJunkAfterListSupVerExpectedVersionListSupVerExpectedVersionNoMutualTlsVersionChExpectedCiphersuiteInapporiateFallbackCiphersuite0NoMutualCompressionChJunkAfterEndNoMutualCertificateTypeSCertTypeExpectedListSCertTypeExpectedFormatKeyShareNotAllowedInTls12ChRenegotiationAttackDetectedShRenegotiationAttackDetectedTls13RequresKeyShareKeySharesNotSubsetSupGrpClientVulernableTHSClientVulernableRenegoChDoubleRestartCkeExpectedPublicFinishedMacFailedCantExtractSpki(CertificateError)InterlockCertificate(CertificateValidationError)NullExtractorBadTls12MsSizeKeyblockTooSmallMasterblockTooSmallCantCreateProtectorCantCreateDeprotectorTextOutputTooSmallConcatenationNotUtf8InterlockNontrivalKexWrongClientRandomLength(usize)WrongServerRandomLength(usize)TlsExtractorsRequireEmsOrTls13Tls13ExporterLabelTooLongTls13ExporterContextTooLongTls13ExporterOutputTooLongCantLockSessionCvSignatureFailedSctRequiredForCertCantParseIssuerCertificate(CertificateError)CkeWrongNumberOfShares(usize)HrrExpectedVersionHrrBogusTlsVersion(u16)HrrTlsVersionUnsupported(u16)HrrExpectedExtensionKindHrrExpectedExtensionPayload(u16)HrrDuplicateExtension(u16)HrrExpectedKeyShareHrrExpectedCookieHrrUnsupportedExtension(u16, u16)HrrExpectedExtensionsHrrExpectedEndHrrBogusRestartHrrAlreadyHadShare(u16)HrrBogusCurve(u16)AlpnExpectedLength1AlpnExpectedLength2AlpnEmptyProtocolAlpnRawLengthTooBigAlpnExpectedLen1Val(usize, usize)AlpnExpectedLen2Val(usize, usize)AlpnNonUtf8ProtocolAlpnBogusProtocolSKeyShareExpectedGroupSKeyShareExpectedKexKeyShareGroupDisagreement(u16, u16)KeyShareBogusGroup(u16)ShExpectedVersionShSsl3NotSupportedShTls10NotSupportedShTls11NotSupportedShBogusVersion(u16)ShUnsupportedTlsVersion(u16)ShExpectedRandomShBadRandomLengthTls13DowngradeAttackShExpectedSessionIdShBogusCiphersuite(u16)ShExpectedCiphersuiteShBadTls12Cipher(u16)ShBadTls13Cipher(u16)ShExpectedCompressionShBogusCompression(u8)ShExpectedExtensionKindShExpectedExtensionPayload(u16)ShDuplicateExtension(u16)SctExpectedListSctExpectedEntrySCertTypeExpectedFormatEShUnsupportedExtension(u16, u16)ShJunkAfterExtension(u16)ShExpectedExtensionsShExpectedEndTls13NoKeyShareServerVulernableTHSServerVulernableRenegoCertRequiresOcspCertRequiresSctSkeExpectedCurveTypeSkeBogusCurveType(u8)SkeExpectedNamedcurveSkeBogusCurve(u16)SkeExpectedPublicSkeCurveError(u16, KeyAgreementError)SkeExpectedAlgorithmSkeBogusSignatureAlgo(u16)SkeExpectedSignatureSkeExpectedEndSkeSignatureFailedServerKeyNotWhitelistedEECertParseError(CertificateError)SCertTypeBogusType(u8)ServerChainEmptyEeCertificateEmptyCsBogusOcspType(u8)CantExtractSct12(OcspValidationError)OcspIncompleteCertificateChainInvalidOcsp(OcspValidationError)CrExpectedContextCrContextMustBeEmptyCantExtractSct13(OcspValidationError)CertificateValidationFailed(CertificateValidationError)CertExpectCertListCertExpectIntermeidateCertIntermediateEmptyCertExpectedContextCertExpectExtensionKindCertExpectExtionPayload(u16)CertDuplicateExtension(u16)OcspExpectedTypeOcspExpectedContentSCertTypeInNotFirstCertCertUnsupportedExtension(u16)CertJunkAfterExtension(u16)CertExpectedExtensionsCvExpectedAlgorithmCvBogusSignatureAlgo(u16)CvExpectedSignatureCvExpectedEndCsExpectedStatusTypeCsExpectedResponseCsEmptyResponseEeExpectedExtensionKindEeExpectedExtensionPayload(u16)EeDuplicateExtension(u16)EeUnsupportedExtension(u16)EeJunkAfterExtension(u16)EeExpectedExtensionsEeExpectedEndDeprotectionFailedAppdataBeforeFinishedAppdataAfterEofRecvAppdataRQueueEofMaxRecordScratchOverflowTxHandshakeMsgOver16MB(usize)FragmentTooBig(usize)InvalidFragRtype(u8)PacketLengthOutOfWhack(usize)RecordTooBig(usize)NoPacketToProcessShortreadRecordHeaderRxRecordTooBig(usize)PacketLengthToSendOutOfWhack(usize)ProtectingPaceketFailedTxQueueWriteFailed(QueueWriteError)NormalCloseProtectingPaceketFailedZReadDiscardFailed(QueueReadError)ValidationHandshakeDoneNoRecordExpectedUnknownIoExceptionTxQueueEofTxQueueDiscard(QueueReadError)RxQueueEofWQueueDataNotReadStreamReadOverflowNoTrustAnchorsCantLockAcmeChallengesInternalConnectionTlsSessionErrorIntErrNoIssuerSerialSctValidationFailed(SctError)
Methods

impl TlsFailure

fn get_alert_num(&self) -> u8

Get TLS alert number that corresponds to the failure.

Note: Some failures return alert #0, which is CLOSE_NOTIFY.

fn get_explanation(&self) -> Cow<'static, str>

Get explanation text for an error.