Function btls::futures::create_future

pub fn create_future<T: Sized + Send>() -> (FutureSender<T>, FutureReceiver<T>)

Create a future value.

This function creates a (sender, receiver) pair for a future value of type T.

Futures:

You don't need to understand futures with this library unless you are changing the way certificate lookup/signing works.

Futures model computation that happens asynchronously, and whose final value may not be known yet. When the value is known, the future settles, and the value of computation can then be extracted

The computation invoked can also be synchronous: In that case, the value settles during invocation of the computation.

Because the values are stored, they need to be Sized and because those can cross threads, the values need to be Send. Because no synchronous mutation is possible, Sync is not needed.

In this implementation of futures, one can:

Create an unsettled future by using this function. It also returns object that can settle the future (and be consumed in process), by using the .settle() method.

Create an immediately settled future by using the FutureReceiver::from function (from From trait).

Check if future has settled by using the .settled() value on the receiver.

When the future has settled, read the value using .read() value on receiver, consuming the receiver.

Return value:

The future sender.

The future receiver.