Struct btls::futures::FutureReceiver

pub struct FutureReceiver<T: Sized + Send>(_);

A future receiver.

This forms the receive half of future value.

If created via the From trait, the value settles immediately to the value FutureReceiver is constructed from.

For more information about futures, see create_future.

Methods

impl<T: Sized + Send> FutureReceiver<T>

fn settled_cb(&mut self, cb: Box<FutureCallback<T> + Send>) -> bool

Set function to call when the future settles.

Parameters:

self: The receiver to use.

cb: The callback to use.

Return value:

True on success, false if future has already settled.

fn settled(&self) -> bool

Has the value settled yet?

Parameters:

self: The receiver to use.

True if value has settled, false if not.

fn read(self) -> Result<T, FutureReceiver<T>>

Read the settled value.

This consumes the future receiver and reads the settled value.

Parameters:

self: The receiver to use.

Returns:

The value.

Failures:

Fails with the receiver wrapped in Err() if the future hasn't settled yet.