Function btls::jose_jws::sign_jws

pub fn sign_jws<K: KeyPair>(prot_header: &JsonNode,
                            payload: &[u8],
                            keypair: &K,
                            algorithm: u16)
                            -> FutureReceiver<Result<SignedJws, ()>>

Sign a JWS and return to future resolving to the signature.

If the prot_header parameter does not already contain a alg parameter, that is automatically inserted.

Parameters:

prot_header: The JWS protected header. Must be a dictionary.

payload: The payload to sign.

keypair: The keypair to sign with.

algorithm: The algorithm to use. One of the SIG_* constants (TLS signature scheme numbers).

Return value:

Future resolving to the signature when available.