Module btls::server_certificates

Types related to server certificate store.

This module contains the various types related to server certificate store. This includes:

The base trait for certificate selection: CertificateLookup.

The base trait for signable certificate: CertificateSigner.

Store of signable certificates loaded from files: LocalServerCertificateStore.

Store of signable certificates loaded from a directory (with automatic managment): CertificateDirectory.

It also contains the certificate-related types: Certificate, CertificateChain, OcspResponse, StapledSct and StapledSctList, plus SIGALGO_* constants.

Structs

Certificate
A single certificate

CertificateChain
A certificate chain.

CertificateDirectory
Certificate store loaded from a directory

CertificateLookupCriteria
Lookup criteria for certificates.

LocalServerCertificateHandle
Handle for local certificate.

LocalServerCertificateStore
A certificate lookup that uses local certificates and associated private keys.

OcspResponse
A single OCSP response.

StapledSct
A single stapled SCT.

StapledSctList
A list of stapled SCTs

Constants

SIGALGO_ECDSA_MASK
The mask for ECDSA-type signature algorithms.

SIGALGO_ECDSA_SHA256
ECDSA P-256 with SHA-256 hash.

SIGALGO_ECDSA_SHA384
ECDSA P-384 with SHA-384 hash.

SIGALGO_ECDSA_SHA512
ECDSA P-512 with SHA-512 hash (not actually supported).

SIGALGO_ED25519
Ed25519

SIGALGO_ED448
Ed448

SIGALGO_RSA_MASK
The mask for RSA-type signature algorithms.

SIGALGO_RSA_PKCS1_SHA256
RSA PKCS #1 v1.5 with SHA-256 hash.

SIGALGO_RSA_PKCS1_SHA384
RSA PKCS #1 v1.5 with SHA-384 hash.

SIGALGO_RSA_PKCS1_SHA512
RSA PKCS #1 v1.5 with SHA-512 hash.

SIGALGO_RSA_PSS_SHA256
RSA PSS with SHA-256 hash.

SIGALGO_RSA_PSS_SHA384
RSA PSS with SHA-384 hash.

SIGALGO_RSA_PSS_SHA512
RSA PSS with SHA-512 hash.

SIGALGO_TLS13_MASK
The mask for TLS 1.3 OK signature algorithms.

Traits

CertificateLookup
Certificate lookup.

CertificateSigner
A certificate chain, plus signer for the associated private key.