Struct btls::server_certificates::CertificateDirectory

pub struct CertificateDirectory { /* fields omitted */ }

Certificate store loaded from a directory

Upon constructing the object:

Files with name *.key are scanned. Found ones are passed to LocalKeyPair for loading. At least 1 successfully loaded key must exist.

Files with name *.crt are scanned. Those that are symbolic links that point to directory are loaded as an certificate.

Periodically, files with name *.crt are rescanned.

If existing *.crt file disappears, the corresponding certificate is deleted.

If new *.crt file that is symbolic link to directory appears, it is loaded.

If existing *.crt file changes destination of symbolic link, the old certificate is unloaded and new one is loaded in its place.

If ocsp file within certificate directory changes, the OCSP response is updated.

If new *.sct file within certificate directory appears, it is loaded as SCT response.

The certificate directories have the following files:

certificate: The certificate itself. Can only have one certificate. Required.

issuer: The issuer certificate. Can have multiple certificates. Required.

ocsp: OCSP staple response. Optional

*.sct: SCT responses.

Periodically, files with name '*.ctlog' are rescanned.

If a new file with '*.ctlog' extension appears, it is loaded as a trusted log.

The file needs to contain 2 or 3 lines,

First line is log name

second being base64 (not base64url!) encoding of the SPKI-format log public key.

Optional third line being <oid>/<hash> (only needed for v2-capable, a.k.a. 6962bis logs)

<oid> is textual-form OID of the log.

<hash> is the hash function, currently only sha256 is supported.

Notes:

All symbolic links to directories in the same directory can be replaced by regular files holding just the name of directory.

Due to race conditions, only replace symbolic links with symbolic links and regular file references with regular file references while this code is running.

This struct implements trait CertificateLookup, so it can be used as certificate lookup.

Yes, really: The ::new() function is all there is to this class.

Methods

impl CertificateDirectory

fn new<L: Logging + Send + Sync + 'static, P: AsRef<Path>>(path: P,
log: L)
-> Result<CertificateDirectory, Cow<'static, str>>

Create a new store that loads certificates from specified directory.

Parameters:

path: The path to directory to load.

log: The logger to use for certificate logging messages.