Struct btls::server_certificates::LocalServerCertificateHandle

pub struct LocalServerCertificateHandle { /* fields omitted */ }

Handle for local certificate.

This handle is used for referencing local certificates, for adding or updating staple data.

Methods

impl LocalServerCertificateHandle

fn clear_ocsp_staple(&mut self) -> Result<(), Cow<'static, str>>

Clear OCSP stapling for certificate.

Parameters:

self: The certificate to update.

Failures:

If operation fails, it returns cause of failure as string.

Notes:

OCSP multishot is cleared as well.

fn set_ocsp_staple<R: IoRead>(&mut self,
staple: &mut R,
name: &str)
-> Result<(), Cow<'static, str>>

Add or update OCSP stapling for certificate.

Parameters:

self: The certificate to update.

staple: The OCSP staple data to associate.

name: Name of OCSP staple to use in error messages.

Failures:

If operation fails, it returns cause of failure as string.

Notes:

The data is in raw ASN.1 form.

Notes:

OCSP multishot is cleared as well.

fn set_ocsp_staple_file<P: AsRef<Path>>(&mut self,
staple: P)
-> Result<(), Cow<'static, str>>

Add or update OCSP stapling for certificate from file.

Parameters:

self: The certificate to update.

staple: The name of file to load OCSP staple data from.

Failures:

If operation fails, it returns cause of failure as string.

Notes:

The data is in raw ASN.1 form.

fn set_ocsp_staple_file_multishot<P: AsRef<Path>>(&mut self, staple: P)

Add or update OCSP stapling for certificate from file, with auto-reload.

This method differs from .set_ocsp_staple_file() in that the passed file is periodically reread as new OCSP stapling data (the first read attempt will occur synchronously with the call).

Parameters:

self: The certificate to update.

staple: The name of file to load OCSP staple data from.

Notes:

The data is in raw ASN.1 form.

fn get_ocsp_last_multishot_error(&self) -> Result<(), (u64, Cow<'static, str>)>

Get error from last failed OCSP multishot operation.

Parameters:

self: The certificate to query.

Failures:

If the last OCSP multishot operation failed, returns the sequence number (first error is 1) of failure since the last .set_ocsp_staple_file_multishot() plus cause of failure as string.

Notes:

This gets cleared to no error on any successful OCSP multishot operation.

fn add_sct_staple<R: IoRead>(&mut self,
staple: &mut R,
name: &str)
-> Result<(), Cow<'static, str>>

Add or update SCT stapling for certificate.

Parameters:

self: The certificate to update.

staple: The SCT staple data to associate.

name: Name of staple to use in error messages.

Failures:

If operation fails, it returns cause of failure as string.

Notes:

The data is in the raw SerializedSct format.

Call this (or .add_sct_staple_file()) multiple times if you have multiple SCTs to send.

fn add_sct_staple_file<P: AsRef<Path>>(&mut self,
staple: P)
-> Result<(), Cow<'static, str>>

Add or update SCT stapling for certificate from file.

Parameters:

self: The certificate to update.

staple: The name of file to load SCT staple data from.

Failures:

If operation fails, it returns cause of failure as string.

Notes:

The data is in the raw SerializedSct format.

Call this (or .add_sct_staple()) multiple times if you have multiple SCTs to send.

fn delete(&self)

Delete the certificate from store.

Parameters:

self: The certificate to delete.