Struct btls::server_certificates::LocalServerCertificateStore

pub struct LocalServerCertificateStore(_, _);

A certificate lookup that uses local certificates and associated private keys.

This structure acts as Certificate lookup (CertificateLookup) that uses the added private keys and certificates to resolve certificate requests.

The private keys are added by .add_privkey() method.

The certificates (after you have added the corresponding private key, can be added using .add() and .add_file() methods.

If you need OCSP or SCT stapling, save the handle resulting from adding the certificate to variable, and call its methods to set the OCSP or SCT stapling.

Methods

impl LocalServerCertificateStore

fn remote_to_local_handle(&self,
handle: &LocalServerCertificateHandle)
-> Option<LocalServerCertificateHandle>

Get local handle for remote handle.

Parameters:

self: The store to get the handle in.

handle: The remote handle.

Returns:

If the handle is also valid in the local store, the handle in this store wrapped in Some().

Otherwise, None.

Note:

Only meaningful if one of the stores is cloned from the other (or if both stores are cloned from the same store).

fn deep_clone(&self) -> LocalServerCertificateStore

Deep-clone the certifcate store. The new store will be duplicate of the old one.

Parameters:

self: The store to clone.

Returns:

The cloned store.

Notes:

Any OCSP multishots are cloned too.

The certificates are shared between the stores. However, any removals or inserts only affect store thos are done in.

Any keys added also only affect the store that is done in.

fn new<L: Logging + Send + Sync + 'static>(log: L)
-> LocalServerCertificateStore

Create a new certificate store.

Parameters:

log: The log to use.

Returns:

New certificate store.

fn add_privkey<K: KeyPair + Send + Sync + 'static>(&mut self,
privkey: K,
name: &str)
-> Result<(), Cow<'static, str>>

Add a private key into store.

Parameters:

self: The certificate lookup to add private key to.

privkey: The private key.

name: The name of key to use in error messages.

Failures:

If operation fails, it returns cause of failure as string.

Notes:

This variant adds arbitrary instance of KeyPair trait, so you can use it to add private keys that the library or application using it never see.

fn add<R: IoRead>(&mut self,
chain: &mut R,
name: &str)
-> Result<LocalServerCertificateHandle, Cow<'static, str>>

Add a certificate to store.

Parameters:

self: The certificate lookup to add certificate to.

chain: The certificate chain.

name: The name of chain to use in error messages.

Return value:

The handle for the certificate. This is used to provode/update stapling data.

Failures:

If operation fails, it returns cause of failure as string.

Notes:

The private key corresponding to the first certificate has to be added before.

If the certificate is not self-signed, the second certificate in chain must sign the first.

This can be used to add certificates from slices, since &[u8]: Read (but note the &mut qualifier on the Read type).

Currently DER and PEM formats are supported (multiple certificates are just concatenated).

fn add_file<P: AsRef<Path>>(&mut self,
chain: P)
-> Result<LocalServerCertificateHandle, Cow<'static, str>>

Load a certificate from file.

Parameters:

self: The certificate lookup to add certificate to.

chain: The filename of certificate chain file to load.

Return value:

The handle for the certificate. This is used to provode/update stapling data.

Failures:

If operation fails, it returns cause of failure as string.

Notes:

The private key corresponding to the first certificate has to be added before.

If the certificate is not self-signed, the second certificate in chain must sign the first.

Currently DER and PEM formats are supported (multiple certificates are just concatenated).

fn add_trusted_log(&mut self, log: &TrustedLog)

Add a trusted CT log.

Parameters:

self: The jar to manipulate.

log: The trusted log to add.

Notes:

Trusted logs have to be added for dedicated SCT staples, in order to sanity-check them.