Trait btls::server_certificates::CertificateLookup

pub trait CertificateLookup {
    fn lookup(&self,
              conditions: &CertificateLookupCriteria)
              -> FutureReceiver<Result<Box<CertificateSigner + Send>, ()>>;
    fn clone_self(&self) -> Box<CertificateLookup + Send>;
}

Certificate lookup.

This trait is for objects that perform certificate lookup (select the certificate used).

The implementations of this trait provoded by the library are:

LocalServerCertificateStore: Load key and certificate files and use those.

You don't have to understand the internals of this trait unless you are changing the way certificate lookup or certificate signing works.

Required Methods

fn lookup(&self,
conditions: &CertificateLookupCriteria)
-> FutureReceiver<Result<Box<CertificateSigner + Send>, ()>>

Look up a certificate.

Parameters:

self: The certificate lookup.

`criteria': The criteria to use in selecting certificate.

Returns:

Future with:

Certificate signer.

Or on failure, ().

Notes:

Because certificate selection can involve contacting an external signer, the return value is a future and not a direct value. If the future has not settled, the handshake will be paused until it is.

fn clone_self(&self) -> Box<CertificateLookup + Send>

Clone this object.

Parameters:

self: The certificate lookup.

Returns:

Clone of the lookup.

Notes:

The reason this exists instead of just requiring Clone is that the latter is not possible if CertificateLookup is to be object-safe.