Trait btls::server_certificates::CertificateSigner

pub trait CertificateSigner {
    fn sign(&self, data: &[u8]) -> FutureReceiver<Result<(u16, Vec<u8>), ()>>;
    fn is_ecdsa(&self) -> bool;
    fn get_certificate(&self) -> CertificateChain;

    fn get_ocsp(&self) -> OcspResponse { ... }
    fn get_scts(&self) -> StapledSctList { ... }
}

A certificate chain, plus signer for the associated private key.

This trait is for objects that represent the results of certificate selection. It provodes signing capability with certificate private key and the certificate info (along with assoicated info).

You don't have to understand the internals of this trait unless you are changing the way certificate lookup or certificate signing works.

Required Methods

fn sign(&self, data: &[u8]) -> FutureReceiver<Result<(u16, Vec<u8>), ()>>

Sign a message.

Parameters:

self: The signer.

data: The raw data to sign.

Returns:

A future for:

The TLS SignatureScheme value.

The actual signature.

Or on failure, ().

Notes:

Because signing can involve contacting an external signer, the return value is a future and not a direct value. If the future has not settled, the handshake will be paused until it is.

fn is_ecdsa(&self) -> bool

Is ECDSA (also EdDSA)?

Parameters:

self: The signer.

Returns:

true if signature scheme is ECDSA or EdDSA, false otherwise.

fn get_certificate(&self) -> CertificateChain

Get the certificate chain.

Parameters:

self: The signer.

Returns:

The associated certificate chain.

Provided Methods

fn get_ocsp(&self) -> OcspResponse

Get OCSP staple for certificate.

Parameters:

self: The signer.

Returns:

The associated OCSP stapling response. Empty if none.

fn get_scts(&self) -> StapledSctList

Get SCT staples for certificate

Parameters:

self: The signer.

Returns:

The associated SCT stapling response.

Notes:

If list of SCTs is empty, no SCT staple gets sent.