Module btls::server_keys

Types related to server keys.

This module contains the various types related to server keys. This includes:

The base trait for server keys: KeyPair.

Key pairs loaded from files, or file-like things: LocalKeyPair.

It also contains auxillary type SubjectPublicKeyInfo.

Structs

EccExported

GnutlsDatum

GnutlsDatumMalloc

GnutlsX509PrivkeyWrap

LocalKeyPair
Key pair that is implemented by signing using a private key.

SubjectPublicKeyInfo
A SubjectPublicKeyInfo.

SupportedSchemes
Supported schemes.

Enums

KeyLoadingError
Error for local key pair

KeyReadError
Error reading local key pair

ModuleError
Error from module.

Traits

KeyPair
A key pair (public key and private key).