Struct btls::server_keys::LocalKeyPair

pub struct LocalKeyPair(_);

Key pair that is implemented by signing using a private key.

This implements the KeyPair trait.

Obviously, given the nature of this kind of signer, the private key is contained in the memory space of the process, and thus can be leaked with any sort of memory disclosure vulernability (not necressarily in the TLS implementation).

Methods

impl LocalKeyPair

fn new<R1: IoRead>(privkey: &mut R1,
name: &str)
-> Result<LocalKeyPair, KeyReadError>

Create a keypair out of Read trait value.

Parameters:

privkey: The private key.

name: Name to use in errors.

Return value:

The keypair.

Failures:

If operation fails, it returns cause of failure as string.

Notes:

This can be used to load keys from buffers since &[u8]: Read (but note the extra &mut qualifier)

Only DER format is currently supported.

fn new_file<P1: AsRef<Path>>(privkey: P1) -> Result<LocalKeyPair, KeyReadError>

Create a keypair out of file.

Parameters:

privkey: The filename of private key file to load.

Return value:

The keypair.

Failures:

If operation fails, it returns cause of failure as string.

Notes:

Only DER format is currently supported.