Trait btls::server_keys::KeyPair

pub trait KeyPair {
    fn sign(&self,
            data: &[u8],
            algorithm: u16)
            -> FutureReceiver<Result<Vec<u8>, ()>>;
    fn get_schemes(&self) -> SupportedSchemes;
    fn get_public_key(&self) -> SubjectPublicKeyInfo;
}

A key pair (public key and private key).

This trait is for objects that represent a keypair. It provodes signing capability with a raw private key and the associated public key.

Required Methods

fn sign(&self,
data: &[u8],
algorithm: u16)
-> FutureReceiver<Result<Vec<u8>, ()>>

Sign a message.

Parameters:

self: The signer.

data: The raw data to sign.

algorithm: The TLS algorithm code to sign with. The following are known:

0x0401 => RSA-PKCS1 /w SHA-256.

0x0501 => RSA-PKCS1 /w SHA-384.

0x0601 => RSA-PKCS1 /w SHA-512.

0x0403 => ECDSA /w SHA-256.

0x0503 => ECDSA /w SHA-384.

0x0603 => ECDSA /w SHA-512.

0x0804 => RSA-PSS /w SHA-256.

0x0805 => RSA-PSS /w SHA-384.

0x0806 => RSA-PSS /w SHA-512.

0x0807 => Ed25519.

0x0808 => Ed448.

Returns:

A future for:

The the algorithm and actual signature, wrapped in Ok().

Or on failure, Err(()).

Notes:

Because signing can involve contacting an external signer, the return value is a future and not a direct value. If the future has not settled, the handshake will be paused until it is.

fn get_schemes(&self) -> SupportedSchemes

Get list of TLS signature schemes this keypair is capable of signing with.

Parameters:

self: The signer.

Return value:

Array of TLS signature scheme values supported.

fn get_public_key(&self) -> SubjectPublicKeyInfo

Get the public key.

Parameters:

self: The signer.

Returns:

The public key in X.509 SubjectPublicKeyInfo format (with the initial SEQUENCE).