Struct btls::ClientConfiguration

pub struct ClientConfiguration { /* fields omitted */ }

Client configuration.

This structure contains various parts of the client configuration.

Setting trust anchors and creating configuration object (required).

You can set up trust anchors and create the configuration like:

let mut talist = Vec::new();
talist.push(TrustAnchor::from_certificate_file("ca1.crt")?);
talist.push(TrustAnchor::from_certificate_file("ca2.crt")?);
let mut config = ClientConfiguration::from(&talist[..]);

If you want to "burn" the key into executable image:

let mut talist = Vec::new();
let cert = include_bytes!("../certificate/test3.der");
talist.push(TrustAnchor::from_certificate(&mut &cert[..])?);
let mut config = ClientConfiguration::from(&talist[..]);

Alternatively, if you are using host-specific pinning with trust anchors, you can also use it as follows:

let mut config = ClientConfiguration::new();

This is handy if you are e.g. implementing DANE (RFC6698) validation (but don't use it for HPKP (RFC7469), see below).

Setting supported ALPNs (optional):

You can set up supported ALPNs like follows:

config.add_alpn("h2");
config.add_alpn("http/1.1");

Setting require EMS (optional):

You can set EMS to be required as follows:

config.set_flags(0, FLAGS0_REQUIRE_EMS);

This is useful if you are going to use exporters, since those won't work without EMS.

Note: Negotiating TLS 1.3 is taken as sufficient to meet the EMS requirement.

Setting TLS version range (optional):

Enabling TLS 1.3 support:

config.set_flags(0, FLAGS0_ENABLE_TLS13);

Requiring TLS 1.3

config.set_flags(0, FLAGS0_ENABLE_TLS13);
config.clear_flags(0, FLAGS0_ENABLE_TLS12);

Certificate Transparency (optional):

Adding a log:

let ct_log = TrustedLog{...};
config.add_ct_log(&ct_log);

Requiring Certificate Transparency.

config.set_flags(0, FLAGS0_REQUIRE_CT);

Making session

When you have it all set up, you can get client session using .make_session().

let mut session = config.make_session(hostname);

Or if you want to specify host-specific pins (mandatory if you used .new()):

const SHAHASH1: [u8; 32] = [66;32/*...*/];
const SHAHASH2: [u8; 32] = [77;32/*...*/];
let pin1 = HostSpecificPin::trust_server_key_by_sha256(&SHAHASH1, false, true);
let pin2 = HostSpecificPin::trust_server_key_by_sha256(&SHAHASH2, false, true);
let pins = [pin1, pin2];
let mut session = config.make_session_pinned(hostname, &pins[..]);

This is useful for implementing DANE (RFC6698) or HPKP (RFC7469). Or if you want to "just test" without "valid" CA certificate.

Methods

impl ClientConfiguration

fn new() -> ClientConfiguration

Create a client configuration with empty trust anchor list.

Note that in order to successfully create a session out of this, you need to use .make_session_pinned() specifying at least one pin with TA flag set (trust_server_key_by_sha256 and from_tlsa_raw (when first byte is 2 or 3) constructors both set the TA flag).

Return value:

The new configuration.

impl ClientConfiguration

fn add_alpn(&mut self, alpn: &str)

Add supported ALPN.

Parameters:

self: The configuration to modify.

alpn: The ALPN to add.

Notes:

The ALPNs are given in order of decreasing perference.

fn set_flags(&mut self, page: usize, mask: u64)

Set flags.

Parameters:

self: The configuration to modify.

page: The page number to set flags on.

mask: The mask of flags to set.

fn clear_flags(&mut self, page: usize, mask: u64)

Clear flags.

Parameters:

self: The configuration to modify.

page: The page number to set flags on.

mask: The mask of flags to clear.

fn config_flags<F>(&mut self, config: &str, error_cb: F) where F: FnMut(ConfigFlagsError)

Set/Clear flags according to config string.

Parameters:

self: The configuration to modify.

config: The config string, a comma-separated string. Each word may be prefixed with '+' to enable (default) or '-' or '!' to disable.

error_cb: Callback to call about unknown flags.

fn set_debug(&mut self, mask: u64, log: Box<Logging + Send + Sync>)

Set debugging mode.

Parameters:

self: The configuration to modify.

mask: Mask of debugging events (DEBUG_*) to log.

log: The log to log the events to.

fn add_ct_log(&mut self, log: &TrustedLog)

Add a trusted certificate transparency log.

Parameters:

self: The configuration to modify.

log: The log to add.

fn blacklist(&mut self, spkihash: &[u8])

Add certificate blacklist entry.

Parameters:

self: The configuration to modify.

spkihash: The SHA-256 hash of the SPKI to blacklist. Always 32 octets.

fn disallow_server_rpk(&mut self)

Disallow raw SPKI for server authentication.

Calling this causes the client to disallow server sending a raw public key, even if client has host- specific pins that are for server SPKI.

This is needed if the raw public keys are provoded by untrusted parties and server sending bogus responses for unknown names is a problem. If the application isn't a browser, this is unlikely to be a problem.

Note that this does not disable matching against server key. The pins continue to be valid.

Parameters:

self: The configuration to modify.

fn make_session(&self, host: &str) -> Result<ClientSession, TlsFailure>

Make a new client session.

Parameters:

self: The configuration to use.

host: The hostname to use (will be signaled as SNI).

Returns:

New client session.

Failures:

On error returns error string.

In practicular, fails if there are no trust anchors set up.

fn make_session_pinned(&self,
host: &str,
pins: &[HostSpecificPin])
-> Result<ClientSession, TlsFailure>

Make a new client session with host-specific pins.

This is handy for DANE (RFC6698) and HPKP (RFC7469).

Parameters:

self: The configuration to use.

host: The hostname to use (will be signaled as SNI).

pins: The pins to use.

Returns:

New client session.

Failures:

On error returns error string.

Unlike .make_session(), this only fails without trust anchors if there are no pins with TA flag set.

fn make_session_insecure_nocheck(&self) -> Result<ClientSession, TlsFailure>

Make a client session WITH NO CHECKING WHATSOEVER.

Establishes session that does absolutely NO CHECKING on what the server is, other than the server key possession check. Without additional checks (e.g. signing TLS Exporter values and verifying the signatures) this is TRIVIALLY INSECURE. If nobody hijacks such connection, it is just because nobody tried.

Restrictions:

The following restrictions apply to these connections:

SNI is disabled, and will not be sent.

RPK support indication is forcibly enabled.

The "EMS support required" is forcibly enabled (i.e. the server MUST support either EMS or TLS 1.3, or the connection will fail).

get_server_names() never succeeds.

Parameters:

self: The configuration to use.

Returns:

New client session.

Failures:

On error returns error string.