Struct btls::ClientSession

pub struct ClientSession(_);

A client TLS session.

This is the client end of TLS session.

Use .make_session() or .make_session_pinned() of ClientConfiguration to create it.

For operations supported, see trait Session.

If the session is cloned, both sessions refer to the same underlying TLS connection. This can be used to refer to the same session from multiple threads at once.