Struct btls::ServerConfiguration

pub struct ServerConfiguration { /* fields omitted */ }

Server configuration.

This structure contains various parts of the server configuration.

Creating a configuration:

You can create the configuration object as:

Setting certificate store and creating configuration (required).

You can use CertificateDirectory as the certificate lookup (unless you need something like "keyless SSL" or "LURK SKI"), like this:

let mut cert_store = CertificateDirectory::new(cert_dir_path, logger);
let mut config = ServerConfiguration::from(&cert_store);

Setting supported ALPNs (optional):

You can set up supported ALPNs like follows:

config.add_alpn("h2");
config.add_alpn("http/1.1");

Note that if there are supported ALPNs, and client hello contains the ALPN extension, the client hello will be rejected if there is no overlap in ALPN values.

Setting require EMS (optional):

You can set EMS to be required as follows:

config.set_flags(0, FLAGS0_REQUIRE_EMS);

This is useful if you are going to use exporters, since those won't work without EMS.

Note: If the negotiatied TLS version is 1.3, that is interpretted as sufficient to pass the EMS requirement, even if EMS is deprecated in TLS 1.3.

Setting TLS version range (optional):

Enabling TLS 1.3 support:

config.set_flags(0, FLAGS0_ENABLE_TLS13);

Requiring TLS 1.3.

config.set_flags(0, FLAGS0_ENABLE_TLS13);
config.clear_flags(0, FLAGS0_ENABLE_TLS12);

Setting ACME challenge directory (strongly recommended to let admin configure this if binding to port 443

for an extended time).

config.set_acme_dir(path_from_config);

Making session

When you have it all set up, you can get server session using .make_session().

Methods

impl ServerConfiguration

fn set_acme_dir<P: AsRef<Path>>(&mut self, path: P)

Set ACME challenge path.

This is STRONGLY RECOMMENDED to be admin-configurable if you bind to port 443.

Parameters:

self: The configuration to modify.

path: The path to set.

fn add_alpn(&mut self, alpn: &str)

Add supported ALPN.

Parameters:

self: The configuration to modify.

alpn: The ALPN to add.

Notes:

The ALPNs are given in order of decreasing perference.

fn set_flags(&mut self, page: usize, mask: u64)

Set flags.

Parameters:

self: The configuration to modify.

page: The page number to set flags on.

mask: The mask of flags to set.

fn clear_flags(&mut self, page: usize, mask: u64)

Clear flags.

Parameters:

self: The configuration to modify.

page: The page number to set flags on.

mask: The mask of flags to clear.

fn config_flags<F>(&mut self, config: &str, error_cb: F) where F: FnMut(ConfigFlagsError)

Set/Clear flags according to config string.

Parameters:

self: The configuration to modify.

config: The config string, a comma-separated string. Each word may be prefixed with '+' to enable (default) or '-' or '!' to disable.

error_cb: Callback to call about unknown flags. Passed the index (0-based) that triggered the warning, a flag telling if setting was tried to be enabled and a flag telling the setting is unknown

fn create_acme_challenge(&mut self,
challenge: &str,
response: Option<&str>)
-> Result<(), TlsFailure>

Create an ACME challenge response endpoint.

Parameters:

self: The settings.

challenge: Challenge to respond to.

response: Optional response to challenge.

Failures:

Error string.

Notes:

If handshake with matching challenge is accepted, that session will immediately disconnect upon reaching showtime state.

fn delete_acme_challenge(&mut self, challenge: &str) -> Result<(), TlsFailure>

Delete an ACME challenge response endpoint.

Parameters:

self: The settings.

challenge: Challenge to delete.

Failures:

Error string.

fn set_debug(&mut self, mask: u64, log: Box<Logging + Send + Sync>)

Set debugging mode.

Parameters:

self: The configuration to modify.

mask: Mask of debugging events (DEBUG_*) to log.

log: The log to log the events to.

fn set_default_server_name(&mut self, name: &str)

Set default server name.

This server name is used if client does not indicate one via server_name extension.

Parameters:

self: The configuration to modify.

name: The default server name.

fn make_session(&self) -> Result<ServerSession, TlsFailure>

Create a server session with these settings.

Parameters:

self: The settings.

Return value.

The session.

Failures:

Error string.