Module btls::test

Internal testing stuff.

All of this is totally unstable.

Reexports

pub use server_keys::*;

Structs

AcmeCertificate
ACME challenge certificate.

AnyAsn1Tag
Match any ASN.1 Tag.

Asn1Tlv
Parsed ASN.1 DER TLV

Base64Decoder

Builder
Thread configuration. Provides detailed control over the properties and behavior of new threads.

DevNull

DhfOutput

DhfPublicKey

DummyDiffieHellmanKey

File
A reference to an open file on the filesystem.

HashInput

HashOutput

HsBuffer

OcspDumpedScts
Information about SCTs dumped from OCSP response.

OcspSanityCheckResults
Structure containing results from basic OCSP sanity check.

OcspValidationResult
Results of OCSP validation.

ParsedCertificate
Parsed X.509 certificate

Path
A slice of a path (akin to str).

PathBuf
An owned, mutable path (akin to String).

Queue

SecBuffer
A Heap buffer that is wiped upon drop.

SecStackBuffer
A stack buffer that is wiped upon drop.

SexprParseStream

SinkMarker
A position marker in sink

SliceMarker
A saved position.

Source
Source that provodes data out of byte slice.

TimeNow

TrustAnchorValue
A trust anchor

Enums

AcmeCertificateCreateError
Error from creating ACME certifcate.

Asn1Error
Error from reading ASN.1 DER TLV.

Asn1Tag
ASN.1 tag

CborNode

CertificateError
Error in parsing certificate.

CertificateExtensionError
Error parsing certificate extension.

CertificateTimestampError
Error parsing a timestamp in certificate.

CertificateToplevelError
Error reading certificate top level.

CertificateValidationError
Error in validating certificate.

CertificateValidityError
Error parsing certificate validity.

CertificateVersionError
Error in certificate version.

ChecksumFunction

DecodingError
Error in decoding.

Dhf

HashFunction

HsBufferError
Error in handshake buffer.

JsonNode

KeyAgreementError
Error in key agreement.

KeyAlgorithm
Signature algorithm key is for.

KeyExchangeType

Length

MajorCertProblem
Variant for the TLS certificate error classification.

NameConstraintsError
Errro in parsing name constraints.

NameListError
Error in parsing name list.

OcspExtensionsError
Error in processing OCSP extensions.

OcspResponseError
Error in OCSP Single Response

OcspValidationError
Error in validating OCSP response.

ProtectorType

QueueReadError
Error from queue read operation.

QueueWriteError
Error from queue write operation.

ResponseStatusError
Bad response_status values

RngError
Error in random number generator.

SctError
Error parsing/validating SCT.

SexprNode

SignatureAlgorithm
Signature algorithm.

SignatureType

TimestampParseError
Error in parsing timestamp.

Constants

ASN1_BIT_STRING
The tag value for ASN.1 BIT STRING type

ASN1_BMP_STRING
The tag value for ASN.1 BMP STRING type

ASN1_BOOLEAN
The tag value for ASN.1 BOOLEAN type

ASN1_CHARACTER_STRING
The tag value for ASN.1 CHARACTER STRING type

ASN1_CTX0_C
The tag value for ASN.1 [0] constructed

ASN1_CTX0_P
The tag value for ASN.1 [0] primitive

ASN1_CTX1_C
The tag value for ASN.1 [1] constructed

ASN1_CTX1_P
The tag value for ASN.1 [1] primitive

ASN1_CTX2_C
The tag value for ASN.1 [2] constructed

ASN1_CTX2_P
The tag value for ASN.1 [2] primitive

ASN1_CTX3_C
The tag value for ASN.1 [3] constructed

ASN1_CTX3_P
The tag value for ASN.1 [3] primitive

ASN1_EMBEDDED_PDV
The tag value for ASN.1 EMBEDDED PDV type

ASN1_ENUMERATED
The tag value for ASN.1 ENUMERATED type

ASN1_GENERALIZED_TIME
The tag value for ASN.1 GENERALIZED TIME type

ASN1_GENERAL_STRING
The tag value for ASN.1 GENERAL STRING type

ASN1_GRAPHICS_STRING
The tag value for ASN.1 GRAPHICS STRING type

ASN1_IA5_STRING
The tag value for ASN.1 IA5 STRING type

ASN1_INSTANCE_OF
The tag value for ASN.1 INSTANCE_OF type

ASN1_INTEGER
The tag value for ASN.1 INTEGER type

ASN1_ISO646_STRING
The tag value for ASN.1 ISO646 STRING type

ASN1_NULL
The tag value for ASN.1 NULL type

ASN1_NUMERIC_STRING
The tag value for ASN.1 NUMERIC STRING type

ASN1_OBJECT_DESCRIPTOR
The tag value for ASN.1 OBJECT DESCRIPTOR type

ASN1_OCTET_STRING
The tag value for ASN.1 OCTET STRING type

ASN1_OID
The tag value for ASN.1 OBJECT IDENTIFIER type

ASN1_PRINTABLE_STRING
The tag value for ASN.1 PRINTABLE STRING type

ASN1_REAL
The tag value for ASN.1 REAL type

ASN1_RELATIVE_OID
The tag value for ASN.1 UTF-8 STRING type

ASN1_SEQUENCE
The tag value for ASN.1 SEQUENCE type

ASN1_SET
The tag value for ASN.1 SET type

ASN1_TELEX_STRING
The tag value for ASN.1 TELEX STRING type

ASN1_UNIVERSAL_STRING
The tag value for ASN.1 UNIVERSAL STRING type

ASN1_UTC_TIME
The tag value for ASN.1 UTC TIME type

ASN1_UTF8_STRING
The tag value for ASN.1 UTF-8 STRING type

ASN1_VIDEO_STRING
The tag value for VIDEO STRING type

CBOR_FALSE

CBOR_NULL

CBOR_TRUE

CBOR_UNDEFINED

CFLAG_MUST_CT
Flag for TLS feature 18: MUST-CT

CFLAG_MUST_EMS
Flag for TLS feature 23: MUST-EMS.

CFLAG_MUST_RENEGO
Flag for TLS feature 65281: MUST-RENEGO

CFLAG_MUST_STAPLE
Flag for TLS feature 5: MUST-STAPLE.

CFLAG_MUST_TLS13
Flag for TLS feature 40: MUST-TLS13.

CFLAG_UNKNOWN
Flag for unknown TLS feature.

KEYFORMAT_ECDSA

KEYFORMAT_ED25519

KEYFORMAT_ED448

KEYFORMAT_RSA

MAX_HASH_OUTPUT

SIG_ECDSA_SHA256
ECDSA signature with SHA-256.

SIG_ECDSA_SHA384
ECDSA signature with SHA-384.

SIG_ECDSA_SHA512
ECDSA signature with SHA-512.

SIG_ED25519
Ed25519 signature.

SIG_ED448
Ed448 signature.

SIG_RSA_PKCS1_SHA256
RSA PKCS #1 v1.5 signature with SHA-256.

SIG_RSA_PKCS1_SHA384
RSA PKCS #1 v1.5 signature with SHA-384.

SIG_RSA_PKCS1_SHA512
RSA PKCS #1 v1.5 signature with SHA-512.

SIG_RSA_PSS_SHA256
RSA PSS signature with SHA-256.

SIG_RSA_PSS_SHA384
RSA PSS signature with SHA-384.

SIG_RSA_PSS_SHA512
RSA PSS signature with SHA-512.

VERIFY_FLAG_ALLOW_SHA1
Flag: Allow SHA1 signatures

VERIFY_FLAG_NO_RSA_PKCS1
Flag: TLS 1.3, disallow RSA-PKCS1

Statics

CURVE_NSA_P256
OID for NSA P-256 curve.

CURVE_NSA_P384
OID for NSA P-384 curve.

ECDSA_KEY_OID
OID for ECDSA keys.

ECDSA_SHA256_OID
OID (including headers) for ECDSA signature with SHA-256.

ECDSA_SHA384_OID
OID (including headers) for ECDSA signature with SHA-384.

ECDSA_SHA512_OID
OID (including headers) for ECDSA signature with SHA-512.

ED25519_OID
OID for Ed25519 algorithm.

ED448_OID
OID for Ed448 algorithm.

LONG_ED25519_OID
OID for Ed25519 algorithm (with header).

LONG_ED448_OID
OID for Ed448 algorithm (with header).

RSA_KEY_OID
OID for RSA keys.

RSA_PKCS1_SHA256_OID
OID&Parameters for RSA-PKCS1v1.5 signature with SHA-256.

RSA_PKCS1_SHA384_OID
OID&Parameters for RSA-PKCS1v1.5 signature with SHA-384.

RSA_PKCS1_SHA512_OID
OID&Parameters for RSA-PKCS1v1.5 signature with SHA-512.

RSA_PSS_KEY_OID
OID for RSA (PSS only) keys.

RSA_PSS_SHA256_OID
OID&Parameters for RSA-PSS signature with SHA-256, MGF1-SHA256 and 32-byte salt.

RSA_PSS_SHA384_OID
OID&Parameters for RSA-PSS signature with SHA-384, MGF1-SHA384 and 48-byte salt.

RSA_PSS_SHA512_OID
OID&Parameters for RSA-PSS signature with SHA-512, MGF1-SHA512 and 64-byte salt.

SHA1_INSECURE_OID
OID for SHA-1 hash (unsafe!)

SHA256_OID
OID for SHA-256 hash

SHA384_OID
OID for SHA-384 hash

SHA512_OID
OID for SHA-512 hash

X509_CSR_EXT_REQ
OID for X.509 CSR request extensions.

X509_EKU_ANY
OID for X.509 AnyEKU EKU.

X509_EKU_OCSP
OID for X.509 OCSP responder EKU.

X509_EKU_TLS_SERVER
OID for X.509 TLS server authentication EKU.

X509_EXT_BASIC_CONSTRAINTS
OID for X.509 BasicConstraints extension.

X509_EXT_EKU
OID for X.509 ExtendedKeyUage extension.

X509_EXT_KEY_USAGE
OID for X.509 KeyUsage extension.

X509_EXT_NAME_CONSTRAINTS
OID for X.509 NameConstraints extension.

X509_EXT_SAN
OID for X.509 SubjectAlternativeName extension.

X509_EXT_SCT
OID for X.509 Signed Certificate Timestamp extension.

X509_EXT_TLS_FEATURE
OID for X.509 TLS Feature extension.

X509_OCSP_BASIC_RESPONSE
OID for X.509 OCSP BasicResponse.

X509_OCSP_SCT
OID for X.509 OCSP SCT.

X509_SUBJECT_COMMON_NAME
OID for X.509 Subject CommonName.

X509_SUBJECT_SERIAL_NUMBER
OID for X.509 Subject SerialNumber.

Traits

Asn1TagMatch
Trait used to cast Asn1Tag matchers to actual values.

Buffer
Trait for "Behaves like a buffer".

DiffieHellmanKey

HashFunctionContext

RangeTrait
Range types for non-panicing slicing

Sink
Sink that consumes raw byte data.

SlicingExt
Extension trait implementing non-panicing subslicing

TimeInterface

Functions

aes_gcm_is_hw

check_sig_algorithm_supported
Is specified signature algorithm supported?

check_sig_algorithm_supported_sha1
Is specified signature algorithm supported (with SHA-1 allowed)?

checksum

convert_key_from_cwk

convert_key_from_jwk

convert_key_from_pem

convert_key_from_sexpr

decode_pem

decode_ts_for_user
Decode timestamp to human-readable format.

extract_key_algorithm
Extract the algorithm for key.

extract_spki
Extract the SPKI (SubjectPublicKeyInfo) from certificate

generate_ed25519_pubkey

generate_ed448_pubkey

interpret_timestamp
Interpret an ASN.1 timestamp

is_executable

is_single_asn1_structure
Is the slice a single ASN.1 structure or not?

is_socket

looks_like_pem

name_in_set2
Check that specified reference name is in specified set (also considering wildcards).

read_dir
Returns an iterator over the entries within a directory.

sanity_check_certificate_chain
Sanity-check a certificate chain. The following checks are done:

sanity_check_certificate_chain_load
Do Load-time sanity-check a certificate chain. Does the same checks as sanity_check_certificate_chain, except the following: - The signature algorithm check is unconditionally done. - Reference SAN check is never done. - There is no callback for postcheck. - Of feature flags, only unknown ones are checked. - Killist is not checked. - The public key is checked against the given one.

secure_random

sleep
Puts the current thread to sleep for the specified amount of time.

slice_eq_ct

stderr
Constructs a new handle to the standard error of the current process.

trans_asn1_oid
Translate ASN.1 OID into textual representation.

validate_chain

validate_ee_spki
Validate EE SPKI against host pins.

validate_scts
Validate a list of SCTs.

validate_signature
Validate a signature under specified key.

wipe_buffer
Wipe a buffer

Type Definitions

ChecksumOutput