Enum btls::test::OcspValidationError

pub enum OcspValidationError {
    SCertParse(CertificateError),
    RCertParse(CertificateError),
    BasicResponse(Asn1Error),
    ResponseSignatureError,
    ResponderSignatureError,
    ResponderNotYetValid(i64),
    ResponderExpired(i64),
    ResponderNotValidForUse,
    ResponderNotIssuedByParent,
    MultipleResponderCertificates,
    ResponderCert(Asn1Error),
    ResponderChainO(Asn1Error),
    ResponderChainI(Asn1Error),
    BrTbs(Asn1Error),
    BrSignature(Asn1Error),
    BrSignatureAlgorithm(Asn1Error),
    OcspResponseStatus(Asn1Error),
    OcspResponseBytesO(Asn1Error),
    OcspResponseBytesI(Asn1Error),
    Toplevel(Asn1Error),
    OcspResponseType(Asn1Error),
    OcspNotBasicResponse,
    BadResponseStatus(ResponseStatusError),
    OcspResponseData(Asn1Error),
    BrSignatureSanity,
    UnknownSignatureAlgo,
    ResponderId(Asn1Error),
    BadResponderId,
    ProducedAt(Asn1Error),
    ProducedAtParse(TimestampParseError),
    ProducedAtInFuture(i64),
    SingleResponses(Asn1Error),
    SingleResponse(usize, OcspResponseError),
    NoSingleResponses,
    Extensions(Asn1Error),
    ExtractScts(OcspExtensionsError),
    CertificateRevoked,
    // some variants omitted
}

Error in validating OCSP response.

Variants
SCertParse(CertificateError)

Error in parsing subject certificate.

This error occurs if parsing of the subject certificate fails.

The parameter is more detailed parsing error.

RCertParse(CertificateError)

Error in parsing responder certificate.

This error occurs if parsing of the responder certificate fails.

The parameter is more detailed parsing error.

BasicResponse(Asn1Error)

Error in reading BasicResponse

This error occurs if reading of the BasicResponse fails.

The parameter is more detailed reading error.

ResponseSignatureError

OCSP response signature does not verify.

This error occurs if the signature in OCSP response does not verify.

ResponderSignatureError

OCPS responder certificate signature does not verify.

This error occurs if the signature in OCSP responder certificate does not verify.

ResponderNotYetValid(i64)

Responder certificate is not yet valid.

This error occurs if the responder certificate sent is not yet valid.

The parameter is the timestamp when the validity of the responder certificate would begin.

ResponderExpired(i64)

Responder certificate has expired.

This error occurs if the responder certificate sent has expired.

The parameter is the timestamp when the validity of the responder certificate ended.

ResponderNotValidForUse

Responder certificate not valid EKU for OCSP responder.

This error occurs if the responder certificate does not have valid ExtendedKeyUsage for OCSP.

ResponderNotIssuedByParent

Responder certificate not issued by the issuer certificate.

This error occurs if the responder certificate is not issued by the same issuer as the certificate being checked.

MultipleResponderCertificates

Multiple responder certificates were sent.

This error occurs if OCSP response contains multiple responder certificates.

ResponderCert(Asn1Error)

Error reading responder certificate element.

This error occurs if reading the responder certificate fails.

The parameter is more detailed reading error.

ResponderChainO(Asn1Error)

Error reading responder chain element (outer).

This error occurs if reading the outer element of responder certificate chain fails.

The parameter is more detailed reading error.

ResponderChainI(Asn1Error)

Error reading responder chain element (inner).

This error occurs if reading the inner element of responder certificate chain fails.

The parameter is more detailed reading error.

BrTbs(Asn1Error)

Error reading TBS element of Basic Response.

This error occurs if reading the to-be-signed data element of Basic Response fails.

The parameter is more detailed reading error.

BrSignature(Asn1Error)

Error reading signature element of Basic Response.

This error occurs if reading the signature element of Basic Response fails.

The parameter is more detailed reading error.

BrSignatureAlgorithm(Asn1Error)

Error reading signature algorithm element of Basic Response.

This error occurs if reading the signature algorithm element of Basic Response fails.

The parameter is more detailed reading error.

OcspResponseStatus(Asn1Error)

Error reading response status element.

This error occurs if reading the response status fails.

The parameter is more detailed reading error.

OcspResponseBytesO(Asn1Error)

Error reading response bytes element (outer).

This error occurs if reading the outer element of response bytes fails.

The parameter is more detailed reading error.

OcspResponseBytesI(Asn1Error)

Error reading response bytes element (inner).

This error occurs if reading the inner element of response bytes fails.

The parameter is more detailed reading error.

Toplevel(Asn1Error)

Error reading top-level of OCSP response.

This error occurs if reading the toplevel of the OCSP response fails.

The parameter is more detailed reading error.

OcspResponseType(Asn1Error)

Error reading OCSP response type element.

This error occurs if reading the OCSP response type fails.

The parameter is more detailed reading error.

OcspNotBasicResponse

OCSP response is not Basic Response.

This error occurs if the OCSP response is not a Basic Response.

BadResponseStatus(ResponseStatusError)

Response status is bad

This error occurs if the response status is bad (i.e. not OK).

The parameter gives more detailed response status.

OcspResponseData(Asn1Error)

Error in reading OCSP response data.

This error occurs if reading of the OCSP response data fails.

The parameter is more detailed reading error.

BrSignatureSanity

Basic Response signature fails sanity check.

This error occurs if the Basic Response signature is clearly out of whack.

UnknownSignatureAlgo

Unknown signature algorithm.

This error occurs if the signature algorithm used is bad.

ResponderId(Asn1Error)

Error in reading responder ID.

This error occurs if reading of the OCSP responder ID fails.

The parameter is more detailed reading error.

BadResponderId

Responder ID is of bad type.

This error occurs if the responder ID is of unknown type.

ProducedAt(Asn1Error)

Error in reading OCSP ProducedAt.

This error occurs if the ProducedAt field of OCSP response can't be read.

The parameter is more detailed reading error.

ProducedAtParse(TimestampParseError)

Error in parsing OCSP ProducedAt into timestamp.

This error occurs if the ProducedAt field of OCSP response can't be parsed.

The parameter is more detailed parsing error.

ProducedAtInFuture(i64)

The OCSP ProducedAt is in the future.

This error occurs if the OCSP ProducedAt timestammp is in the future.

SingleResponses(Asn1Error)

Error in reading Single Responses list.

This error occurs if the Single Responses list field of OCSP response can't be read.

The parameter is more detailed reading error.

SingleResponse(usize, OcspResponseError)

Error in parsing Single Responses entry.

This error occurs if parsing of OCSP Single Response fails.

The 1st parameter is number of the esponse (starting from 0) that triggered the parsing error.

The 2nd parameter is more detailed parsing error.

NoSingleResponses

There are no Single Responses in OCSP response.

This error occurs if the Single Responses list in OCSP response is empty.

Extensions(Asn1Error)

Error in reading OCSP extensions.

This error occurs if reading of the toplevel of OCSP extensions fails.

The parameter is more detailed reading error.

ExtractScts(OcspExtensionsError)

Error in extracting SCTs from OCSP response.

This error occcurs if parsing SCTs from OCSP response fails.

The parameter is more detailed parsing error.

CertificateRevoked

Certificate has been REVOKED.