Enum btls::test::TimestampParseError

pub enum TimestampParseError {
    InvalidDate(i64, usize, i64),
    InvalidTime(u32, u32, u32),
    UnrecognizedFormat,
    BadUtcTime,
    BadGeneralizedTime,
    // some variants omitted
}

Error in parsing timestamp.

Variants
InvalidDate(i64, usize, i64)

The date is invalid.

This error occurs if a date is invalid (invalid day in month, invalid month or invalid year).

The 1st parameter is the year.

The 2nd parameter is the month.

The 3rd parameter is the day.

InvalidTime(u32, u32, u32)

The time is invalid.

This error occurs if time is invalid (invalid hour number, invalid minute number or invalid second number).

The 1st parameter is the hour.

The 2nd parameter is the minute.

The 3rd parameter is the second.

UnrecognizedFormat

Unrecognized time format.

This error occurs if timestamp isn't either UTCTime nor GeneralizedTime.

BadUtcTime

UTCTime notation is bad.

This error occurs if UTCTime value can not be parsed into its components.

BadGeneralizedTime

GeneralizedTime notation is bad.

This error occurs if GeneralizedTime value can not be parsed into its components.