Function btls::test::sanity_check_certificate_chain

pub fn sanity_check_certificate_chain<'b, T, F, P>(_eecert: &'b [u8],
                                                   _icerts: &[&'b [u8]],
                                                   reference: Option<&str>,
                                                   time_now: &TimeInterface,
                                                   strict_algo: bool,
                                                   feature_flags: u32,
                                                   callback: F,
                                                   cbparam: P,
                                                   killist: &BTreeMap<[u8; 32], ()>,
                                                   need_ocsp_flag: &mut bool,
                                                   need_sct_flag: &mut bool)
                                                   -> Result<T, CertificateValidationError> where F: FnMut(&ParsedCertificate<'b>, &[ParsedCertificate<'b>], &mut BTreeMap<&'b [u8], &'b [u8]>, &BTreeMap<&'b [u8], ()>, P, &BTreeMap<[u8; 32], ()>, i64, u32) -> Result<T, CertificateValidationError>

Sanity-check a certificate chain. The following checks are done:

Eecert must have SAN for the reference (if any).

The two signature algorithm slots match up.

No certificate may be expired or not yet valid.

eecert must have use_tls_serv and !is_ca

Every non-eecert must have (use_tls_serv | use_any) and is_ca.

Every certificate (except the first) must sign some previous certificate.

Every EE SAN satisfies all certificate name constraints.

Subject name must be key for subject public key (also checked versus initial values of keys_map)

EE cert must have unique public key.

The first CA cert must have issued the EE cert.

Path length constraints are satisfied.

If strict_algo is set, every signature is with approved signature method.

Also, keys_map is updated and issuers_seen is populated with certificate issuers seen.

The parameters passed to callback are in order: - Parsed EE certificate - Parsed intermediate certificates. - Borrow of keys_map. - Set of issuer names seen. - Specified argument to sanity_check_certificate_chain. - Expiry time of the certificate - The required flags for certificate.

Parameters:

_eecert: The EE certificate.

_icerts: The intermediate certificates.

reference: The reference name (if any).

time_now: Current time.

strict_algo: If true, strictly check algorithms.

feature_flags: The feature flags that have been met.

callback: The callback to call after validation.

cbparam: The extra parameter to pass to callback.

need_ocsp_flag: Set if certificate requires OCSP.

need_sct_flag: Set if certificate requires SCT.

Return value:

The return value of the callback.

Failures:

If the validation fails, returns an appropriate error.