Struct btls::test::Asn1Tlv

pub struct Asn1Tlv<'a> {
    pub tag: Asn1Tag,
    pub value: Source<'a>,
    pub raw_p: &'a [u8],
    pub outer: &'a [u8],
}

Parsed ASN.1 DER TLV

Fields
tag: Asn1Tag

The tag.

value: Source<'a>

Source with payload.

raw_p: &'a [u8]

Raw payload.

outer: &'a [u8]

Raw outer (payload, including tag and length).

Methods

impl<'a> Asn1Tlv<'a>

fn call<T: Sized, F>(&self, closure: F) -> T where F: FnMut(&mut Source<'a>) -> T

Call closure on value (closure(&mut self.value)).

Parameters:

self: The TLV.

closure: The closure to call.

Return value:

Whatever closure returns.

fn fold<T: Sized, E: Sized, F>(&self, initial: T, closure: F) -> Result<T, E> where F: FnMut(&mut Source<'a>, T) -> Result<T, E>

Fold closure on value until end of source is reached.

Parameters:

self: The TLV.

closure: The closure to call.

initial: Initial value of folding.

Returns:

Whatever the value of folded variable at end of iteration.

Failures:

If the closure fails, whatever the error returned.