Struct btls::test::SinkMarker

pub struct SinkMarker(_, _);

A position marker in sink

This is used for subslice lengths and Sink::callback_after.

Methods

impl SinkMarker

fn commit<S: Sink>(self, sink: &mut S) -> Result<(), usize>

Commit the marker

This method writes the length of marked section into the sink.

Parameters:

self: The marker to write. Consumed.

sink: The sink to write to. Must be the same the marker was obtained from.

Failures:

If the section is too large, fails with the size of the section.