Struct btls::test::Source

pub struct Source<'a>(_, _);

Source that provodes data out of byte slice.

This structure reads byte slice as various data types, in order. Handy for parsing all sorts of protocols.

Methods

impl<'a> Source<'a>

fn new<'b>(slice: &'b [u8]) -> Source<'b>

Create a source out of byte slice.

Parameters:

slice: The slice to create the source out of.

Return value:

The newly created source.

fn at_end(&self) -> bool

Is the subslice at end?

Parameters:

self: The source to interrogate.

Return value.

True if at EOF.

False otherwise.

fn as_slice(&self) -> &'a [u8]

Return the source as slice.

Parameters:

self: The source to interrogate.

Return value.

The source data as slice.

fn read_slice<E: Sized + Clone>(&mut self,
len: usize,
err: E)
-> Result<&'a [u8], E>

Read a slice out of source.

Parameters:

self: The source to read from.

len: The length of slice to read.

err: The error to return if read fails.

Return value:

The newly-read slice.

Failures:

If the source does not have enough data to satisfy request, fails with Err(err).

fn read_slice_len<E: Sized + Clone>(&mut self,
length: Length,
err: E)
-> Result<&'a [u8], E>

Read a length and slice out of source.

Parameters:

self: The source to read from.

length: The length notation to use.

err: The error to return if read fails.

Return value:

The newly-read slice.

Failures:

If the source does not have enough data to satisfy request, fails with Err(err).

fn read_remaining(&mut self) -> &'a [u8]

Read everything remaining in source as a slice.

Parameters:

self: The source to read from.

Return value:

The newly-read slice.

fn read_source<E: Sized + Clone>(&mut self,
len: usize,
err: E)
-> Result<Source<'a>, E>

Read a source out of source.

Parameters:

self: The source to read from.

len: The length of new source

err: The error to return if read fails.

Return value:

The newly-read source.

Failures:

If the source does not have enough data to satisfy request, fails with Err(err).

fn read_source_len<E: Sized + Clone>(&mut self,
length: Length,
err: E)
-> Result<Source<'a>, E>

Read a length and source out of source.

Parameters:

self: The source to read from.

length: The length notation to use

err: The error to return if read fails.

Return value:

The newly-read source.

Failures:

If the source does not have enough data to satisfy request, fails with Err(err).

fn read_source_fn<F, T: Sized, E: Sized + Clone>(&mut self,
length: Length,
repeat: bool,
closure: F,
readfail: E,
blankret: T)
-> Result<T, E> where F: FnMut(&mut Source<'a>) -> Result<T, E>

Read a length and source out of source and call closure.

Parameters:

self: The source to read from.

length: The length notation to use

repeat: If true, call closure until it fails or end of source is reached. Return is from last call.

closure: The closure to call.

readfail: Error return to use if reading fails.

Failures:

If the source does not have enough data to satisfy request, fails with Err(readfail).

If closure fails, fails with whatever error closure returned.

fn read_u8<E: Sized + Clone>(&mut self, err: E) -> Result<u8, E>

Read a big-endian u8 out of source.

Parameters:

self: The source to read from.

err: The error to return if read fails.

Return value:

The newly-read integer.

Failures:

If the source does not have enough data to satisfy request, fails with Err(err).

fn read_u16<E: Sized + Clone>(&mut self, err: E) -> Result<u16, E>

Read a big-endian u16 out of source.

Parameters:

self: The source to read from.

err: The error to return if read fails.

Return value:

The newly-read integer.

Failures:

If the source does not have enough data to satisfy request, fails with Err(err).

fn read_u24<E: Sized + Clone>(&mut self, err: E) -> Result<u32, E>

Read a big-endian u24 out of source.

Parameters:

self: The source to read from.

err: The error to return if read fails.

Return value:

The newly-read integer.

Failures:

If the source does not have enough data to satisfy request, fails with Err(err).

fn read_u32<E: Sized + Clone>(&mut self, err: E) -> Result<u32, E>

Read a big-endian u32 out of source.

Parameters:

self: The source to read from.

err: The error to return if read fails.

Return value:

The newly-read integer.

Failures:

If the source does not have enough data to satisfy request, fails with Err(err).

fn read_u64<E: Sized + Clone>(&mut self, err: E) -> Result<u64, E>

Read a big-endian u64 out of source.

Parameters:

self: The source to read from.

err: The error to return if read fails.

Return value:

The newly-read integer.

Failures:

If the source does not have enough data to satisfy request, fails with Err(err).

fn read_asn1_value<Tag: Asn1TagMatch, E, EM>(&mut self,
xtag: Tag,
maperr: EM)
-> Result<Asn1Tlv<'a>, E> where EM: Fn(Asn1Error) -> E

Read an ASN.1 DER TLV with known type out of source.

Only lengths up to 16777215 are read, larger lengths trigger an error.

Parameters:

self: The source to read from.

xtag: The expected tag (AnyAsn1Tag if any tag is to be allowed).

maperr: Closure to map the error on failure.

Return value:

The newly-read ASN.1 TLV.

Failures:

If the source does not have enough data to satisfy request, fails with Asn1Error::Truncated.

If the input is not canonical, fails with Asn1Error::NotDer.

If the length is greater than 16777215, fails with Asn1Error::LengthTooBig.

If the tag is greater than 263-1, fails with Asn1Error::TagTooBig.

If the tag is unexpected, fails with Asn1Error::UnexpectedType.

fn read_asn1_value_opt_fn<T, E, Fc, Fe>(&mut self,
tag: Asn1Tag,
wrong_type: T,
cb: Fc,
errmap: Fe)
-> Result<T, E> where Fc: FnMut(Source<'a>) -> Result<T, E>, Fe: Fn(Asn1Error) -> E

Read an ASN.1 DER TLV out of source and call callback if tag is correct.

Only lengths up to 16777215 are read, larger lengths trigger an error.

Parameters:

self: The source to read from.

tag: The expected tag.

wrong_type: The value to return if wrong type.

cb: The callback to call.

errmap: The error mapper that is passed the errors.

Return value:

Whatever cb returned.

Failures:

If the underlying read fails, whatever errmap returns when passed the appropriate error.

If cb fails, whatever error from it.

fn slice_marker(&self) -> SliceMarker

Get beginning marker for raw slice.

Parameters:

self: The source to read from.

Return value:

The slice marker.

fn position(&self) -> usize

Get the current position in slice.

Parameters:

self: The source to read from.

Return value:

The current position.