Trait btls::test::Sink

pub trait Sink {
    fn write_slice(&mut self, data: &[u8]) -> Result<(), ()>;
    fn _alter(&mut self, ptr: usize, data: u8);
    fn _readback(&self, ptr: usize) -> u8;
    fn written(&self) -> usize;
    fn _pop(&mut self, amount: usize);
    fn callback_after(&self, marker: SinkMarker, cb: &mut FnMut(&[u8]) -> ());

    fn write_u8(&mut self, data: u8) -> Result<(), ()> { ... }
    fn write_u16(&mut self, data: u16) -> Result<(), ()> { ... }
    fn write_u24(&mut self, data: u32) -> Result<(), ()> { ... }
    fn write_u32(&mut self, data: u32) -> Result<(), ()> { ... }
    fn write_u64(&mut self, data: u64) -> Result<(), ()> { ... }
    fn alter_u8(&mut self, ptr: usize, data: u8) -> Result<(), ()> { ... }
    fn alter_u16(&mut self, ptr: usize, data: u16) -> Result<(), ()> { ... }
    fn alter_u24(&mut self, ptr: usize, data: u32) -> Result<(), ()> { ... }
    fn readback_u24(&self, ptr: usize) -> Result<u32, ()> { ... }
    fn marker_posonly(&self) -> SinkMarker { ... }
    fn vector_fn<'a, F>(&'a mut self, length: Length, cb: F) -> Result<(), ()> where Self: Sized, F: FnMut(&mut Self) -> Result<(), ()> { ... }
    fn asn1_fn<'a, F>(&'a mut self, tag: Asn1Tag, cb: F) -> Result<(), ()> where Self: Sized, F: FnMut(&mut Self) -> Result<(), ()> { ... }
}

Sink that consumes raw byte data.

Required Methods

fn write_slice(&mut self, data: &[u8]) -> Result<(), ()>

Append a slice of data into sink.

Parameters:

self: The sink to write to.

data: The data to write.

Failures:

If the write causes the internal capacity of the sink to be exceeded, fails with ().

fn _alter(&mut self, ptr: usize, data: u8)

Alter past-written byte in the sink.

Parameters:

self: The sink to alter the data in.

ptr: Offset to the data to alter.

data: The new value for byte pointed to.

fn _readback(&self, ptr: usize) -> u8

Readback past-written byte in the sink.

Parameters:

self: The sink to readback the data in.

ptr: Offset to the data to read.

Return value:

The readback bytes.

fn written(&self) -> usize

Get number of bytes written so far.

Parameters:

self: The sink to interrogate.

Returns:

The number of bytes written into sink so far.

fn _pop(&mut self, amount: usize)

Pop specified amount of data from the end.

Parameters:

self: The sink to manipulate.

amount: The amount of data to pop.

fn callback_after(&self, marker: SinkMarker, cb: &mut FnMut(&[u8]) -> ())

Get a callback for part after marker.

Parameters:

self: The sink to get data for.

marker: The marker.

cb: The callback.

Provided Methods

fn write_u8(&mut self, data: u8) -> Result<(), ()>

Write a big-endian u8 into sink.

Parameters:

self: The sink to write to.

data: The data to write.

Failures:

If the write causes the internal capacity of the sink to be exceeded, fails with ().

fn write_u16(&mut self, data: u16) -> Result<(), ()>

Write a big-endian u16 into sink.

Parameters:

self: The sink to write to.

data: The data to write.

Failures:

If the write causes the internal capacity of the sink to be exceeded, fails with ().

fn write_u24(&mut self, data: u32) -> Result<(), ()>

Write a big-endian u24 into sink.

Parameters:

self: The sink to write to.

data: The data to write.

Failures:

If the write causes the internal capacity of the sink to be exceeded, fails with ().

fn write_u32(&mut self, data: u32) -> Result<(), ()>

Write a big-endian u32 into sink.

Parameters:

self: The sink to write to.

data: The data to write.

Failures:

If the write causes the internal capacity of the sink to be exceeded, fails with ().

fn write_u64(&mut self, data: u64) -> Result<(), ()>

Write a big-endian u64 into sink.

Parameters:

self: The sink to write to.

data: The data to write.

Failures:

If the write causes the internal capacity of the sink to be exceeded, fails with ().

fn alter_u8(&mut self, ptr: usize, data: u8) -> Result<(), ()>

Alter past-written u8 in the sink.

Parameters:

self: The sink to alter the data in.

ptr: Offset to the data to alter.

data: The new value for data pointed to.

Failures:

If the pointer is out of existing written range, fails with ().

fn alter_u16(&mut self, ptr: usize, data: u16) -> Result<(), ()>

Alter past-written u16 in the sink.

Parameters:

self: The sink to alter the data in.

ptr: Offset to the data to alter.

data: The new value for data pointed to.

Failures:

If the pointer is out of existing written range, fails with ().

fn alter_u24(&mut self, ptr: usize, data: u32) -> Result<(), ()>

Alter past-written u24 in the sink.

Parameters:

self: The sink to alter the data in.

ptr: Offset to the data to alter.

data: The new value for data pointed to.

Failures:

If the pointer is out of existing written range, fails with ().

fn readback_u24(&self, ptr: usize) -> Result<u32, ()>

Readback past-written u24 in the sink.

Parameters:

self: The sink to alter the data in.

ptr: Offset to the data to alter.

Return value:

the value read back.

Failures:

If the pointer is out of existing written range, fails with ().

fn marker_posonly(&self) -> SinkMarker

Get a marker.

Parameters:

self: The sink to get marker for.

fn vector_fn<'a, F>(&'a mut self, length: Length, cb: F) -> Result<(), ()> where Self: Sized, F: FnMut(&mut Self) -> Result<(), ()>

Write size marker of specified size, with data from closure.

Parameters:

self: The sink to write a subslice to.

length: The type of length to write.

cb: The closure to get the data from.

Failures:

If the write causes the internal capacity of the sink to be exceeded, fails with ().

If the write is too large for vector size, fails with ().

fn asn1_fn<'a, F>(&'a mut self, tag: Asn1Tag, cb: F) -> Result<(), ()> where Self: Sized, F: FnMut(&mut Self) -> Result<(), ()>

Write ASN.1 TLV, with data from callback.

Parameters:

self: The sink to write a subslice to.

tag: The ASN.1 tag to use.

cb: The callback to obtain the data from.

Failures:

If the write causes the internal capacity of the sink to be exceeded, fails with ().

If the tag is bigger than 263, fails with ().