Trait btls::test::SlicingExt

pub trait SlicingExt {
    fn slice_np<X: RangeTrait>(&self, r: X) -> Option<&Self>;
    fn slice_np_mut<X: RangeTrait>(&mut self, r: X) -> Option<&mut Self>;
    fn split_at_np<'a>(&'a self, pt: usize) -> Option<(&'a Self, &'a Self)>;

    fn slice_np_len(&self, start: usize, len: usize) -> Option<&Self> { ... }
    fn slice_np_len_mut(&mut self,
                        start: usize,
                        len: usize)
                        -> Option<&mut Self> { ... }
}

Extension trait implementing non-panicing subslicing

This trait provodes method to extract subslice from slice without panicing. Great for extracting slices with indexes of unknown provenance.

This trait is blanket-implemented for all slice types ([T]).

Required Methods

fn slice_np<X: RangeTrait>(&self, r: X) -> Option<&Self>

Extract subslice.

Parameters:

self: The slice to subslice.

r: The range. Can be any of the four usual slice range types.

Return value:

If slicing was successful, Some(x), where x is the slice.

If the range is out of bounds, None.

fn slice_np_mut<X: RangeTrait>(&mut self, r: X) -> Option<&mut Self>

Extract mutable subslice.

Parameters:

self: The slice to subslice.

r: The range. Can be any of the four usual slice range types.

Return value:

If slicing was successful, Some(x), where x is the slice.

If the range is out of bounds, None.

fn split_at_np<'a>(&'a self, pt: usize) -> Option<(&'a Self, &'a Self)>

Split a slice into two.

Parameters:

self: The slice to subslice.

pt: The point to slice at.

Return value:

If slice can be split at given point, Some((a, b)), where a and b are the slices.

If the point is out of range, None.

Provided Methods

fn slice_np_len(&self, start: usize, len: usize) -> Option<&Self>

Extract subslice with specified start and length.

Parameters:

self: The slice to subslice.

start: The start of slice index (0-based).

len: The length of slice.

Return value:

If slicing was successful, Some(x), where x is the slice.

If the range is out of bounds, None.

fn slice_np_len_mut(&mut self, start: usize, len: usize) -> Option<&mut Self>

Extract mutable subslice with specified start and length.

Parameters:

self: The slice to subslice.

start: The start of slice index (0-based).

len: The length of slice.

Return value:

If slicing was successful, Some(x), where x is the slice.

If the range is out of bounds, None.