Trait btls::Session

pub trait Session: IoRead + IoWrite {
    fn set_high_water_mark(&mut self, amount: usize);
    fn set_send_threshold(&mut self, threshold: usize);
    fn bytes_available(&self) -> usize;
    fn bytes_queued(&self) -> usize;
    fn is_eof(&self) -> bool;
    fn wants_read(&self) -> bool;
    fn set_tx_irq(&mut self, handler: Box<TxIrqHandler + Send>);
    fn set_appdata_rx_fn(&mut self,
                         handler: Option<Box<RxAppdataHandler + Send>>);
    fn read_tls<R: IoRead>(&mut self, stream: &mut R) -> IoResult<()>;
    fn submit_tls_record(&mut self, record: &mut [u8]) -> Result<(), TlsFailure>;
    fn write_tls<W: IoWrite>(&mut self, stream: &mut W) -> IoResult<()>;
    fn zerolatency_write(&mut self,
                         output: &mut [u8],
                         input: &[u8])
                         -> Result<(usize, usize), TlsFailure>;
    fn estimate_zerolatency_size(&self, outsize: usize) -> usize;
    fn can_tx_data(&self) -> bool;
    fn handshake_completed(&self) -> bool;
    fn send_eof(&mut self);
    fn aborted_by(&self) -> Option<AbortReason>;
    fn extractor(&self,
                 label: &str,
                 context: Option<&[u8]>,
                 buffer: &mut [u8])
                 -> Result<(), TlsFailure>;
    fn get_alpn(&self) -> Option<Option<Arc<String>>>;
    fn get_server_names(&self) -> Option<Arc<Vec<String>>>;
    fn get_sni(&self) -> Option<Option<Arc<String>>>;
    fn connection_info(&self) -> TlsConnectionInfo;
    fn get_record_size(&self, buffer: &[u8]) -> Option<usize>;
    fn begin_transacted_read<'a>(&self,
                                 buffer: &'a mut [u8])
                                 -> Result<&'a [u8], TlsFailure>;
    fn end_transacted_read(&mut self, len: usize) -> Result<(), TlsFailure>;
}

A TLS session.

This trait models an TLS session.

It is implemented by ClientSession for TLS client, and ServerSession for TLS server.

Plaintext buffering:

For reading, plaintext is buffered. You can use .bytes_available() to query how many bytes are available for reading in buffer (single read produces the amount of bytes requested up to at least that limit).

To prevent excessive read buffering, you can use .set_high_water_mark() to adjust the threshold where the session shuts off reading of the incoming TLS stream.

The read buffer can be accessed by using the impiled Read trait, and its methods (e.g .read()).

For writing, plaintext is also buffered. Now there are no automatic shutoffs, but you can use the .bytes_queued() method to check how many bytes are in the buffer.

The write buffer can be accessed by using the impiled Write trait and its methods (e.g. .write()).

The writes are buffered to 16kB chunks. If you want to output in smaller chunks, use the .flush() method of the Write trait to force output to be immediately sent. Note that this does not work before the handshake has completed to the point that data can be sent.

To end the stream, use .send_eof(). After this is done, no further writes are allowed. On read side, .read() returns Ok(0) forever on EOF. One can also query if read side is at EOF by using .is_eof().

Ciphertext read/write:

Ciphertext read/write happens independently in both directions:

For read side, one can query if session wants to read data from socket by using .wants_read(). Then when the socket is ready, one can use .read_tls() to read from the socket. Exactly one .read() call is performed.

For write side, once there is something to write, the callback registered by .set_tx_irq() is called. Then after callback has arrived, calling .write_tls() performs exactly one .write() on the socket. But do not do the .write_tls() call within the callback! Also note that if there is already a pending write, then calling .set_tx_irq() will call the new irq handler.

One can call .write_tls() even without something to write. In this case, no .write() on socket happens.

Handshake:

When the session is first established, the sides do a handshake. This happens as number of reads and writes during which no data is exchanged. One can query if this has completed by using .handshake_completed().

Data can be queued for sending before the handshake is complete. This data is then sent as soon as possible.

Connection closure:

Upon connection termination, one can obtain the cause of termination from .aborted_by(). The cause can be one of:

MutualEof: Both sides exchanged EOF signals. Connection shut down normally.

ReadError: Connection was terminated by socket read error. The description of error is available as parameter.

WriteError: Connection was terminated by socket write error. The description of error is available as parameter.

SentAlert: Connection was terminated by this side sending a fatal TLS alert. The alert number is included as paramter.

RecvAlert: Connection was terminated by remote side sending a fatal TLS alert. The alert number is included as paramter.

HandshakeError: The connection was terminated by error during handshaking, or due to an error in handshake protocol during running connection. The description of error is available as parameter.

ALPN:

After the handshake is far enough, one can get the ALPN connection uses by using .get_alpn(). This method returns Some(x) if ALPN has been determined. x can be None if no ALPN has been negotiated, or Some(y) where y is the negotiated ALPN.

Server names:

After the handshake is far enough, one can get the list of names server authenticated as by using .get_server_names(). Note that on client side, for security reasons, no results are given until server Finished message has been received and validated (in TLS 1.2, this won't happen until the very last step before the handshake completes, in TLS 1.3 it will happen a bit earlier).

Required Methods

fn set_high_water_mark(&mut self, amount: usize)

Set the receive high water mark.

Sets the receive high water mark of the session. If the number of pending unread bytes is at or above the high water mark, read side will be shut down until read buffer is brought below this value.

Parameters:

self: The session.

amount: New high water mark.

fn set_send_threshold(&mut self, threshold: usize)

Set send threshold.

This sets the send thresohld of the session. If number of pending written bytes is greater than this, the bytes are flushed, even without explicit flush call.

Parameters:

self: The session.

amount: New send threshold (0 enables nodelay mode).

Notes:

If this threshold is increased past the amount of data already buffered, the change is only made effective on the next write. This is to ensure that TX interrupts can never be cleared except by acknowledging them via .write_tls().

fn bytes_available(&self) -> usize

Get the number of bytes currently available for read.

Returns the number of bytes that are waiting in the read buffer. A read is guaranteed to return lesser of this and read buffer size of output.

Parameters:

self: The session.

amount: New high water mark.

Return value:

Number of bytes available.

fn bytes_queued(&self) -> usize

Get the number of bytes currently outstanding for write.

Returns the number of bytes that are waiting in the write buffer (does not count bytes already encrypted for sending).

Parameters:

self: The session.

Return value:

Number of bytes waiting.

fn is_eof(&self) -> bool

Is EOF?

Returns if the read side is at EOF, and no further input will be produced.

Parameters:

self: The session.

Return value:

true if at EOF, false if more input will possibly come.

Notes:

This only covers normal close_notify. It does not cover abnormal termination.

fn wants_read(&self) -> bool

Session wants to read from stream?

This determines if call to read_tls() would read from the stream or not.

Parameters:

self: The session.

Return value:

true if read_tls would read from the stream, false if it would not.

fn set_tx_irq(&mut self, handler: Box<TxIrqHandler + Send>)

Set the TX IRQ handler.

This sets a handler to be called when .write_tls() would write at least one segment (including when `.write_tls() itself does not write all the data it wants to).

If the session wants to write data at the moment this function is called, the specified handler will be called synchronously.,

Parameters:

self: The session.

handler: The new handler.

fn set_appdata_rx_fn(&mut self, handler: Option<Box<RxAppdataHandler + Send>>)

Set the application data RX handler.

This sets a handler to be called when application data record is received. This handler is called from within .read_tls() or .submit_tls_record().

If this handler is set, then the received data is not appended to the RX buffer, so the Read trait methods can't be used when this handler is set.

Parameters:

self: The session.

handler: The new handler. If None, the existing handler is cleared.

fn read_tls<R: IoRead>(&mut self, stream: &mut R) -> IoResult<()>

Read from stream.

Reads TLS stream from specified source.

Parameters:

self: The session.

stream: The stream to read.

Failures:

If there is low-level read error, returns the appropriate error. These errors are not fatal.

If there is fatal error in TLS connection, returns that error as TlsFailure wrapped in IoError.

fn submit_tls_record(&mut self, record: &mut [u8]) -> Result<(), TlsFailure>

Submit TLS record for processing.

Submits TLS record directly for processing, bypassing all per-record buffering. When combined with .set_appdata_rx_fn(), this allows true zero-copy processing of received application data.

Parameters:

self: The session.

record: The record to receive, including the headers. Used as scratchpad during the call.

Failures:

If fatal error occurs in TLS record processing, returns that error.

If connection is not expecting data, fails with TlsFailure::NoRecordExpected. This is not fatal.

fn write_tls<W: IoWrite>(&mut self, stream: &mut W) -> IoResult<()>

Write to stream.

Writes TLS stream to specified sink.

Parameters:

self: The session.

stream: The stream to write.

Failures:

If there is low-level write error, returns the appropriate error. These errors are not fatal.

If there is fatal error in TLS connection, returns that error as TlsFailure wrapped in IoError.

fn zerolatency_write(&mut self,
output: &mut [u8],
input: &[u8])
-> Result<(usize, usize), TlsFailure>

Zero-latency write data into TLS stream.

Directly compies the plaintext data from input buffer to TLS records in output buffer. This call never buffers the read data.

Parameters:

self: The session.

output: The output buffer.

input: The input buffer.

Return value:

Number of bytes that have been written to output buffer.

Number of bytes that have been read from input buffer.

Failures:

Any fatal error in TLS processing is returned as an error.

Notes:

If this function does not run into any fatal TLS error, but does not make any progress either, it returns Ok((0, 0)), that is: Successful, 0 bytes read, 0 bytes written.

fn estimate_zerolatency_size(&self, outsize: usize) -> usize

Get ESTIMATE for zero-latency write maximum size.

This does not take all factors involved in maximum write size into account. As such, it is approximate, but should be correct most of the time, if there is no data in send buffers.

In practicular, the following factors cause it to return 0:

Handshake not finished transmitting.

Any data in write buffer.

Parameters:

self: The session.

outsize: The ouput size.

Returns:

An estimate how much zerolatency_write() would eat output, or 0 if this can't be estimated.

fn can_tx_data(&self) -> bool

Can send data?

Parameters:

self: The session.

Returns:

true if data can be sent, false if not.

Notes:

This flag can be asserted before the handshake completes. However, one should wait until handshake completes handshake_completed(), before actually sending data.

fn handshake_completed(&self) -> bool

Handshake complete?

Parameters:

self: The session.

Returns:

true if the handshake has completed, false if not.

Notes:

One should not send data before completion of the handshake. However, one can write data before that point, and such data can be sent on the wire before the handshake completes.

fn send_eof(&mut self)

Send EOF.

Sends an EOF. After EOF has been sent, no further writes are possible. When the EOF bubbles through to the wire side, it gets translated into close_notify alert.

Parameters:

self: The session.

fn aborted_by(&self) -> Option<AbortReason>

Get error that aborted the TLS connection.

Parameters:

self: The session.

Returns:

If the connection is still up, None.

Otherwise, the error that caused the session to fail.

fn extractor(&self,
label: &str,
context: Option<&[u8]>,
buffer: &mut [u8])
-> Result<(), TlsFailure>

Extract keying material.

Parameters:

self: The session.

label: The label to use.

context: The context to use (None means no context).

buffer: The buffer to store to.

Failures:

If the parameters are invalid, fails with ().

If the handshake isn't complete, fails with ().

If EMS or TLS 1.3 has not been negotiated, fails with ().

fn get_alpn(&self) -> Option<Option<Arc<String>>>

Get the APLN in effect.

Parameters:

self: The session.

Return value:

If the ALPN isn't known yet, None.

If the ALPN is known not to exist, Some(None).

If the ALPN is known to exist, Some(Some(alpn)), where alpn is the alpn.

fn get_server_names(&self) -> Option<Arc<Vec<String>>>

Get the names server authenticated as.

Parameters:

self: The session.

Return value:

If the name list isn't known yet, None.

Otherwise, Some(x), where x is the name list.

fn get_sni(&self) -> Option<Option<Arc<String>>>

Get the SNI client sent.

Note that SNI should only be used for routing, and for protocols that do not have host headers on application level. Do not use it with protocols that do have host headers, e.g. HTTP.

Parameters:

self: The session.

Return value:

If the name isn't known yet, None.

Otherwise, Some(x), where x is the name (might be None if no SNI was sent).

fn connection_info(&self) -> TlsConnectionInfo

Get information about TLS connection.

Return value:

Information about connection.

fn get_record_size(&self, buffer: &[u8]) -> Option<usize>

Get size of first record in buffer.

Parameters:

self: The session to query.

buffer: The buffer to inspect.

Return value:

If the buffer has insufficient contents to even say how large the record is, None.

Otherwise, Some(size), where size is the size of record (including header) in bytes.

fn begin_transacted_read<'a>(&self,
buffer: &'a mut [u8])
-> Result<&'a [u8], TlsFailure>

Begin transactional application data read.

This method reads from unencrypted read buffer of TLS session. The data read is not removed.

Parameters:

self: The session to manipulate.

buffer: Buffer to store the data read to.

Returns:

On success, buffer containing the data (sliced from buffer).

On failure, an error.

Notes:

If there is no data to be read, this just succeeds with an empty buffer.

Do not start multiple transactional reads at once, it won't work.

Example:

let mut buf = [0u8; 16384];
let buf = sess.begin_transactional_read(&mut buf).unwrap();
let wlen = appsock.write(buf).unwrap();
sess.end_transactional_read(wlen).unwrap();

fn end_transacted_read(&mut self, len: usize) -> Result<(), TlsFailure>

End transactional application data read.

This method discards specified amount of data from unencrypted read buffer of TLS session.

Parameters:

self: The session to manipulate.

len: The number of bytes successfully processed.

Returns:

On success, Ok(()).

On failure, an error.

Notes:

See Session::begin_transactional_read() for an example.