Trait btls::TxIrqHandler

pub trait TxIrqHandler {
    fn wants_transmit(&mut self);
}

A TLS Session TX IRQ handler.

This trait models TX IRQ handler, called when the session wants to write something.

Required Methods

fn wants_transmit(&mut self)

The session wants to transmit data.

This method is called when the TLS session wants to transmit data. You should respond by arranging to call session::write_tls() on the session as soon as practical (but not synchronously from the interrupt).

If the call to session::write_tls() does not write all the data it wants to write in one call, this handler will be called from within session::write_tls() to re-arm the transmission.